% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_checks.R
\name{active_annuities_ok}
\alias{active_annuities_ok}
\title{Checking active_annuities dataframe for possible mistakes}
\usage{
active_annuities_ok(active_annuities, last_orig_year)
}
\arguments{
\item{active_annuities}{Dataframe of active annuities, see \link{active_annuities_xmpl}.}

\item{last_orig_year}{Last origin year.}
}
\value{
TRUE if every check is passed, else FALSE.
}
\description{
This function checks this input data for possible mistakes, implausible entries or missing or double rows.
}
\examples{
active_annuities_ok(active_annuities_xmpl, 2023)
}
