% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{add_classes}
\alias{add_classes}
\title{Adding reserve classes to extended_claims_data}
\usage{
add_classes(extended_claims_data, reserve_classes)
}
\arguments{
\item{extended_claims_data}{dataframe after step 5, see details of \code{\link[=prepare_data]{prepare_data()}}.}

\item{reserve_classes}{numeric vector to specify reserve classes at the niveau of index_year, see details of \code{\link[=prepare_data]{prepare_data()}}.}
}
\value{
dataframe with entry and exit reserve classes
}
\description{
The columns \code{Entry_reserve_class} and \code{Exit_reserve_class} can now be derived from \code{Ind_entry_cl_reserve} and \code{Ind_cl_reserve}. \cr \cr

\strong{This function resembles only step 6 of the data preparation process, see description in details of \code{\link[=prepare_data]{prepare_data()}}}
}
\examples{
# this example uses data provided with this package
library(dplyr)
new_data <- minimal_claims_data_xmpl \%>\%
   reduce_data(indices_xmpl, 400000) \%>\%
   add_missing_years(2023) \%>\%
   add_columns() \%>\%
   add_indexed_columns(indices_xmpl) \%>\%
   filter_large_claims(400000, 1989, 3) \%>\%
   add_classes(c(1, 200001, 400001, 700001, 1400001))
head(new_data)
}
