% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{add_columns}
\alias{add_columns}
\title{Add columns}
\usage{
add_columns(extended_claims_data)
}
\arguments{
\item{extended_claims_data}{dataframe after step 2, see details of \code{\link[=prepare_data]{prepare_data()}}.}
}
\value{
dataframe with added columns
}
\description{
This step adds derived columns to \code{claims_data} that are needed for further operations. \cr \cr
\strong{This function resembles only step 3 of the data preparation process, see description in details of \code{\link[=prepare_data]{prepare_data()}}}
}
\examples{
# this example uses data provided with this package
library(dplyr)
minimal_claims_data_xmpl \%>\%
   reduce_data(indices_xmpl, 400000) \%>\%
   add_missing_years(2023) \%>\%
   add_columns()

}
