% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{add_indexed_columns}
\alias{add_indexed_columns}
\title{Add indexed columns}
\usage{
add_indexed_columns(extended_claims_data, indices)
}
\arguments{
\item{extended_claims_data}{dataframe after step 3, see details of \code{\link[=prepare_data]{prepare_data()}}.}

\item{indices}{dataframe, see details of \code{\link[=prepare_data]{prepare_data()}}.}
}
\value{
dataframe with added columns
}
\description{
This step uses the dataframe indices to index every payment and every reserve to the index year. \cr \cr
\strong{This function resembles only step 4 of the data preparation process, see description in details of \code{\link[=prepare_data]{prepare_data()}}}
}
\examples{
# this example uses data provided with this package
library(dplyr)
minimal_claims_data_xmpl \%>\%
   reduce_data(indices_xmpl, 400000) \%>\%
   add_missing_years(2023) \%>\%
   add_columns() \%>\%
   add_indexed_columns(indices_xmpl)

}
