% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{add_missing_years}
\alias{add_missing_years}
\title{Add missing rows}
\usage{
add_missing_years(claims_data, last_orig_year)
}
\arguments{
\item{claims_data}{dataframe after step 1, see details of \code{\link[=prepare_data]{prepare_data()}}.}

\item{last_orig_year}{Integer value for the last origin year which will also be treated as the last calendar year,
usually the calendar year just ended, see details of \code{\link[=prepare_data]{prepare_data()}}.}
}
\value{
dataframe with added rows
}
\description{
\code{claims_data} is requested to only contain rows in which one of \code{Cl_reserve}, \code{An_reserve}, \code{Cl_payment_cal} or \code{An_payment_cal} is not equal to 0. \cr
This step adds rows so that claims_data contains one row per calendar year between the origin year and the last origin year for each claim. \cr \cr
\strong{This function resembles only step 2 of the data preparation process, see description in details of \code{\link[=prepare_data]{prepare_data()}}}
}
\examples{
# this example uses data provided with this package
library(dplyr)
minimal_claims_data_xmpl \%>\%
   reduce_data(indices_xmpl, 400000) \%>\%
   add_missing_years(2023)

}
