% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Excel_Output.R
\name{add_sicr_aggr_worksheet}
\alias{add_sicr_aggr_worksheet}
\title{Create Excel sheet by adding or subtracting other sheets}
\usage{
add_sicr_aggr_worksheet(
  workbook,
  sheetname,
  sheets_to_add = NULL,
  sheets_to_subtract = NULL
)
}
\arguments{
\item{workbook}{Workbook created via \code{openxlsx2::wb_workbook()}.}

\item{sheetname}{Desired sheet name.}

\item{sheets_to_add}{Vector with the names of the sheets that shall be added.}

\item{sheets_to_subtract}{Vector with the names of the sheets that shall be subtracted.}
}
\value{
Workbook including the new sheet.
}
\description{
This function can be used after creating excel worksheets with \code{\link[=add_sicr_worksheet]{add_sicr_worksheet()}} to derive
worksheets as aggregate of these. Addition and subtraction are supported to allow for a net after reinsurance
perspective.
}
\examples{
future_payments <- matrix(0, 35, 250)
colnames(future_payments) <- 2024:2273
future_payments[,1] <- 1e6 # some entries to show functionality

payments <- cbind(all_claims_paid_xmpl, future_payments)

workbook <- openxlsx2::wb_workbook()
workbook <- add_sicr_worksheet(workbook = workbook,
                               sheetname = "add1",
                               payments = payments,
                               reserved = all_claims_reserved_xmpl)
workbook <- add_sicr_worksheet(workbook = workbook,
                               sheetname = "add2",
                               payments = payments,
                               reserved = all_claims_reserved_xmpl)
workbook <- add_sicr_worksheet(workbook = workbook,
                               sheetname = "sub1",
                               payments = payments,
                               reserved = all_claims_reserved_xmpl)

workbook <- add_sicr_aggr_worksheet(workbook = workbook,
                                   sheetname = "Sum_sheet",
                                   sheets_to_add = c("add1", "add2"),
                                   sheets_to_subtract = c("sub1"))

# either open with openxlsx2::wb_open(workbook) or save via openxlsx2::wb_save(workbook)

}
