% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{add_transition_factor}
\alias{add_transition_factor}
\title{adding column Transition_factor to dataframe expanded_indices}
\usage{
add_transition_factor(expanded_indices, index_year)
}
\arguments{
\item{expanded_indices}{dataframe, see details}

\item{index_year}{desired year that the payments are indexed to}
}
\value{
dataframe indices with new column Transition_factor
}
\description{
The transition_factor for year j indexes a payment of year j to the niveau of payments in the fixed index_year.
}
\details{
\strong{expanded_indices} must be a dataframe with columns \code{Calendar_year} and \code{Index_gross} and \code{Index_re}. \cr
\strong{expanded_indices} must contain one row for each historic \code{Calendar_year} that
shall be used plus 250 future calendar years.
\itemize{
\item \code{Index_gross} is the claim payment development from one year to another, e.g. 0.02 for 2\% increase
\item \code{Index_re} is the contractually fixed claim payment development that is to be used in special index clauses that are a common part of longtail xl resinsurance programs. If this column is missing, it will be set to 0.
}
}
\examples{
expanded_indices <- expand_historic_indices(data.frame(Calendar_year = 2015:2023,
                                                       Index_gross = 0.03),
                                            first_orig_year = 2015,
                                            last_orig_year = 2023,
                                            0.03,
                                            0.02)
add_transition_factor(expanded_indices, index_year = 2022)
}
