% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_checks.R
\name{age_shift_ok}
\alias{age_shift_ok}
\title{Checking age_shift dataframe for possible mistakes}
\usage{
age_shift_ok(age_shift)
}
\arguments{
\item{age_shift}{Dataframe, see description of \link{age_shift_xmpl}.}
}
\value{
TRUE if every check is passed, else FALSE.
}
\description{
This function checks this input data for possible mistakes, implausible entries or missing or double rows.
}
\examples{
age_shift_ok(age_shift_xmpl)
}
