% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{age_shift_xmpl}
\alias{age_shift_xmpl}
\title{(unrealistic!) example of an age shift dataframe}
\format{
\subsection{\code{age_shift_xmpl}}{

A data frame with 228 rows and 3 columns:
\describe{
\item{Gender}{"m" or "f" for male or female}
\item{Birth_year}{Birth year}
\item{Age_shift}{Age shift}
}
}
}
\source{
{Made up dataframe to show functionalities of the package.}
}
\usage{
age_shift_xmpl
}
\description{
Dataframe contains one row for each combination of gender and birth year. \cr \cr
As mortality tables are calibrated to a specific birth year, earlier or later birth years must be \code{shifted}. \cr \cr
Example: A shift of -3 for a 50 year old means that the mortality of a 47 year old will be used. \cr \cr
This data belongs to the set of \code{mortality_xmpl}, \code{age_shift_xmpl} and is totally made up to show functionalities of the package. \cr
}
\keyword{datasets}
