% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_large_claims_results.R
\name{aggregate_large_claims_results}
\alias{aggregate_large_claims_results}
\title{Aggregating data and simulation results to payments and reserves matrices}
\usage{
aggregate_large_claims_results(
  sim_result,
  large_claims_list,
  large_extended_claims_data,
  history,
  reinsurance = NULL,
  indices,
  first_orig_year,
  last_orig_year
)
}
\arguments{
\item{sim_result}{Numeric array as a result of \code{\link[=sicr]{sicr()}}.}

\item{large_claims_list}{Dataframe with one row per known large claim generated by \code{\link[=generate_claims_list]{generate_claims_list()}}.}

\item{large_extended_claims_data}{Prepared claims data dataframe, see details of \code{\link[=prepare_data]{prepare_data()}}.}

\item{history}{numeric matrix with one row per claim with historic payments,
generated by \code{\link[=generate_history_per_claim]{generate_history_per_claim()}}.}

\item{reinsurance}{Dataframe with reinsurance information, see details of \code{\link[=xl_cashflow]{xl_cashflow()}} or \link{reinsurance_xmpl}. Default: NULL\cr
Only necessary if reinsurance shall be considered.}

\item{indices}{Dataframe for indexation, see details of \code{\link[=prepare_data]{prepare_data()}}.}

\item{first_orig_year}{First origin year with complete history.}

\item{last_orig_year}{Last origin year.}
}
\value{
List with eight results matrices for large claims. Each matrix consist of one row per
development year between \code{first_orig_year} and \code{last_orig_year}. Matrices with suffix \code{...payments} consist of
one column for each historic calendar year between \code{first_orig_year} and \code{last_orig_year} and 250 predicted
calendar years. Matrices with suffix \code{...reserved} only contain the historic part as no reserves are predicted. \cr
\enumerate{
\item \code{large_claims_payments} with historic and predicted claim payments (e.g. without annuities).
\item \code{large_claims_reserved} with historic claim reserves (without annuities).
\item \code{large_known_annuities_payments} with historic and predicted payments for known annuities.
\item \code{large_annuities_reserved} with historic annuities' reserves.
\item \code{large_future_annuities_payments} with predicted payments for future annuities, the historic part is all 0.
\item \code{large_ibnr_payments} with predicted payments for future annuities, the historic part is all 0.
\item \code{large_ceded_xl_payments} with historic and predicted reinsurance payments for xl treaties.
\item \code{large_ceded_quota_payments} with historic and predicted reinsurance payments for quota share treaties. \cr \cr
}

Note that there are no reserves for future annuities and ibnr. reserved triangles for reinsurance are often not available,
so they are not included here. \cr
As no reserves are predicted (only reserve classes), the index clause in xl treaties can only be an approximation, so
the results of matrix 7 will probably not correspond exactly to existing xl triangles.
}
\description{
This function converts the simulation results and the associated data into payments and reserves matrices.
}
\examples{
# this example uses data provided with this package
extended_claims_data <- prepare_data(claims_data = claims_data_xmpl,
                                     indices = indices_xmpl,
                                     threshold = 400000,
                                     first_orig_year = 1989,
                                     last_orig_year = 2023,
                                     expected_year_of_growing_large = 3,
                                     reserve_classes = c(1, 200001, 400001, 700001, 1400001),
                                     pool_of_annuities = pool_of_annuities_xmpl)

pools <- generate_pools(extended_claims_data = extended_claims_data,
                        reserve_classes = c(1, 200001, 400001, 700001, 1400001),
                        years_for_pools = 2014:2023,
                        start_of_tail = 17,
                        end_of_tail = 50,
                        lower_outlier_limit = -Inf,
                        upper_outlier_limit = Inf,
                        pool_of_annuities = pool_of_annuities_xmpl)

large_claims_list <- generate_claims_list(extended_claims_data = extended_claims_data,
                                          first_orig_year = 1989,
                                          last_orig_year = 2023)

history <- generate_history_per_claim(data = extended_claims_data,
                                      column = "Cl_payment_cal",
                                      first_orig_year = 1989,
                                      last_orig_year = 2023)
# smallest version
result <- sicr(n = 1,
               large_claims_list = large_claims_list,
               first_orig_year = 1989,
               last_orig_year = 2023,
               pools = pools,
               indices = indices_xmpl,
               history = history)

aggregate_large_claims_results(sim_result = result,
                               large_claims_list = large_claims_list,
                               large_extended_claims_data = extended_claims_data,
                               history = history,
                               indices = indices_xmpl,
                               first_orig_year = 1989,
                               last_orig_year = 2023)
}
