% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_annuities_paid_xmpl}
\alias{all_annuities_paid_xmpl}
\title{Example matrix with payments for all annuities (e.g. small + large + special claims)}
\format{
\subsection{\code{all_annuities_paid_xmpl}}{

A numeric matrix with 35 rows and 35 columns that contains all annuities' payments.
}
}
\source{
{ Made up to show package functionalities.}
}
\usage{
all_annuities_paid_xmpl
}
\description{
Example matrix with payments for all annuities (e.g. small + large + special claims)
}
\keyword{datasets}
