% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{apply_index}
\alias{apply_index}
\title{Indexing payments to another year}
\usage{
apply_index(payments, original_cal_years, new_cal_years, indices)
}
\arguments{
\item{payments}{numeric vector of payments to index}

\item{original_cal_years}{vector containing payments' original calendar years}

\item{new_cal_years}{vector containing calendar years to index to}

\item{indices}{dataframe containing transition factors for each element of original_cal_years and new_cal_years, see details}
}
\value{
numeric vector with indexed payments
}
\description{
Indexing payments to another year
}
\details{
The Dataframe indices must contain at least the two columns 'Calendar_year' and 'Transition_factor'. \cr
The transition_factor for year j indexes a payment of year j to the niveau of payments in a fixed index year i.
}
\examples{
# for a constant inflation of 2\% and index year 2022
indices <- data.frame(Calendar_year = 2015:2023,
                      Index_gross = rep(0.02, 9))
indices$Transition_factor <- cumprod(1 + indices$Index_gross) /
                               cumprod(1 + indices$Index_gross)[indices$Calendar_year == 2022]
indices
apply_index(1000, 2022, 2023, indices)
apply_index(payments = rep(1000, 5),
            original_cal_years = 2017:2021,
            new_cal_years = rep(2020, 5),
            indices = indices)
}
