% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{attach_future_annuities}
\alias{attach_future_annuities}
\title{Attach future annuities to extended_claims_data}
\usage{
attach_future_annuities(extended_claims_data, pool_of_annuities = NULL)
}
\arguments{
\item{extended_claims_data}{dataframe after step 6, see details of \code{\link[=prepare_data]{prepare_data()}}.}

\item{pool_of_annuities}{dataframe containing one row for each annuity that has ever been agreed on, see details of \code{\link[=prepare_data]{prepare_data()}}.}
}
\value{
prepared dataframe
}
\description{
The columns \code{New_annuity_1} to \code{New_annuity_5} are added to the dataframe. If new annuities have been agreed on
in the calendar year and the claim of the row, these columns contain the row number of these annuities in the corresponding
dataframe \code{pool_of_annuities}. These columns are needed for the pool as the sampling shall consider claim payments and
new reserve classes as well as new annuities. \cr \cr

The function always adds \strong{five} columns for new annuities which is sufficient for the rare case that five new annuities are agreed on
for one single claim in one calendar year. Each further new annuity in one year will be ignored, which could lead to a slight underestimation
of the best estimate. To avoid this, data may be manually adjusted, for example by moving further annuities to another calendar year or
by aggregating annuities. \cr \cr

\strong{This function resembles only step 7 of the data preparation process, see description in details of \code{\link[=prepare_data]{prepare_data()}}}
}
\examples{
# this example uses data provided with this package
library(dplyr)
new_data <- minimal_claims_data_xmpl \%>\%
   reduce_data(indices_xmpl, 400000) \%>\%
   add_missing_years(2023) \%>\%
   add_columns() \%>\%
   add_indexed_columns(indices_xmpl) \%>\%
   filter_large_claims(400000, 1989, 3) \%>\%
   add_classes(c(1, 200001, 400001, 700001, 1400001)) \%>\%
   attach_future_annuities(minimal_pool_of_annuities_xmpl)
head(new_data)
}
