% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{backtesting_multiyears}
\alias{backtesting_multiyears}
\title{Compares expected and actual payments for more than one year}
\usage{
backtesting_multiyears(
  sim_result_series_old,
  extended_claims_data_new,
  reserve_classes = NULL,
  dev_years = NULL
)
}
\arguments{
\item{sim_result_series_old}{Object created by \code{\link[=sicr_series]{sicr_series()}} containing a series of possible realisations.}

\item{extended_claims_data_new}{Object created by \code{\link[=prepare_data]{prepare_data()}} at least one year later.}

\item{reserve_classes}{Integer vector of desired reserve classes.}

\item{dev_years}{Integer vector of desired development years.}
}
\value{
ggplot2 object
}
\description{
This function compares expected and actual payments for desired reserve classes and development years
on a multi-year basis.
}
\examples{
# this example uses data provided with this package
claims_data_old <- dplyr::filter(claims_data_xmpl,
                                 Calendar_year <= 2020)

historic_indices_old <- dplyr::filter(historic_indices_xmpl,
                                      Calendar_year <= 2020)

indices_old <- expand_historic_indices(historic_indices_old,
                                       first_orig_year = 1989,
                                       last_orig_year = 2020,
                                       index_gross_future = 0.03,
                                       index_re_future = 0.02)

indices_old <- add_transition_factor(indices_old, index_year = 2022)
extended_claims_data <- prepare_data(claims_data = claims_data_old,
                                     indices = indices_old,
                                     threshold = 400000,
                                     first_orig_year = 1989,
                                     last_orig_year = 2020,
                                     expected_year_of_growing_large = 3,
                                     reserve_classes = c(1, 200001, 400001, 700001, 1400001),
                                     pool_of_annuities = pool_of_annuities_xmpl)
pools <- generate_pools(extended_claims_data = extended_claims_data,
                        reserve_classes = c(1, 200001, 400001, 700001, 1400001),
                        years_for_pools = 2014:2020,
                        start_of_tail = 17,
                        end_of_tail = 50,
                        lower_outlier_limit = -Inf,
                        upper_outlier_limit = Inf,
                        pool_of_annuities = pool_of_annuities_xmpl)

large_claims_list <- generate_claims_list(extended_claims_data = extended_claims_data,
                                          first_orig_year = 1989,
                                          last_orig_year = 2020)

history <- generate_history_per_claim(data = extended_claims_data,
                                      column = "Cl_payment_cal",
                                      first_orig_year = 1989,
                                      last_orig_year = 2020)

series <- sicr_series(n = 10,
                      large_claims_list = large_claims_list,
                      first_orig_year = 1989,
                      last_orig_year = 2020,
                      pools = pools,
                      indices = indices_old,
                      history = history)

extended_claims_data_new <- prepare_data(claims_data = claims_data_xmpl,
                                     indices = indices_xmpl,
                                     threshold = 400000,
                                     first_orig_year = 1989,
                                     last_orig_year = 2023,
                                     expected_year_of_growing_large = 3,
                                     reserve_classes = c(1, 200001, 400001, 700001, 1400001),
                                     pool_of_annuities = pool_of_annuities_xmpl)

g <- backtesting_multiyears(sim_result_series_old = series,
                            extended_claims_data_new = extended_claims_data_new,
                            reserve_classes = 3,
                            dev_year = 2:16)
}
