% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_claims.R
\name{best_estimate2cashflow}
\alias{best_estimate2cashflow}
\title{rolling out best estimate vector}
\usage{
best_estimate2cashflow(best_estimate, pattern, last_orig_year)
}
\arguments{
\item{best_estimate}{numeric vector. last entry is for the youngest origin year}

\item{pattern}{numeric vector}

\item{last_orig_year}{integer.}
}
\value{
numeric matrix of future payments per origin year with 250 future calendar years
}
\description{
rolling out the best estimate for the last n origin years with given pattern
}
\examples{
best_estimate2cashflow(c(100, 400, 1000), c(1.7, 1.3, 1.05, 1, 1), 2020)
}
