% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_checks.R
\name{check_column_type}
\alias{check_column_type}
\title{Check column type}
\usage{
check_column_type(column, type)
}
\arguments{
\item{column}{Vector.}

\item{type}{Character. One of \code{integer}, \code{numeric} or \code{character}.}
}
\value{
TRUE if the vector is of specified type.
}
\description{
This is a helper function for data checks. It checks if a vector is of the desired type.
}
\examples{
check_column_type(1:5, "integer")
check_column_type(paste0("test", 1:5), "character")
check_column_type(c(1:4, "notanumber"), "numeric")
}
