% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_checks.R
\name{check_dataframe_structure}
\alias{check_dataframe_structure}
\title{Check structure of a dataframe}
\usage{
check_dataframe_structure(df, structure)
}
\arguments{
\item{df}{Dataframe}

\item{structure}{Dataframe with two columns.
\enumerate{
\item \code{Colname} contains the colnames of \code{df}.
\item \code{Coltype} contains the associated type to check for.
}}
}
\value{
TRUE if \code{df} contains all the columns specified in \code{structure} with the correct types.
}
\description{
This is a helper function for data checks. It checks if the dataframe \code{df} contains all the columns
with names and types specified in \code{structure}.
}
\examples{
df <- data.frame(Col1 = 1:5, Col2 = paste0("test", 1:5))
structure <- data.frame(Colname = c("Col1", "Col2"), Coltype = c("integer", "character"))
check_dataframe_structure(df, structure)
}
