% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_claims.R
\name{cl_factor}
\alias{cl_factor}
\title{chain ladder factor}
\usage{
cl_factor(triangle, dev_year, volume = NULL)
}
\arguments{
\item{triangle}{cumulative claims triangle. An (mxn)-matrix where only the upper left triangle (including diagonal) will be used}

\item{dev_year}{development year to calculate chain ladder factor for. An Integer > 0}

\item{volume}{Number of newest calendar years to be used. Default: NULL,
An Integer > 0 or NULL for no limitation}
}
\value{
A number
}
\description{
calculating the chain-ladder factor for a given development year and claims triangle
}
\examples{
cl_factor(matrix(c(1000, 800, 1200, 1200, 900, 1200, 1300, 900, 1200), 3), 1)
}
