% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{claims_data_xmpl}
\alias{claims_data_xmpl}
\title{Example of a claims data dataframe}
\format{
\subsection{\code{claims_data_xmpl}}{

A data frame with 49000 rows and 7 columns:
\describe{
\item{Claim_id}{Claim ID}
\item{Origin_year}{Origin year of the claim}
\item{Calendar_year}{Calendar year}
\item{Cl_payment_cal}{Claim payments without annuities}
\item{Cl_reserve}{Claim payments without annuities}
\item{An_payment_cal}{Annuities' payments}
\item{An_reserve}{Annuities' reserve}
}
}
}
\source{
{A subset derived by reducing and anonymising an original dataset of an insurer.}
}
\usage{
claims_data_xmpl
}
\description{
Dataframe contains one row for each calendar year per claim where at least one of the
columns \code{Cl_payment_cal}, \code{Cl_reserve}, \code{An_payment_cal}, \code{An_reserve} is not 0. \cr
Data belongs to the set of \code{claims_data_xmpl}, \code{active_annuities_xmpl} and \code{pool_of_annuities_xmpl}.
}
\keyword{datasets}
