% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{convert_single2dev_year_prediction}
\alias{convert_single2dev_year_prediction}
\title{Converting a prediction per single claim into a prediction per development year.}
\usage{
convert_single2dev_year_prediction(
  cashflow,
  claims_orig_years,
  first_orig_year,
  last_orig_year
)
}
\arguments{
\item{cashflow}{matrix with one row per claim and 250 future calendar years}

\item{claims_orig_years}{vector with origin years corresponding to claims in 'cashflow'}

\item{first_orig_year}{desired first origin year}

\item{last_orig_year}{desired last origin year}
}
\value{
sum of payments matrix with origin years as rows and 250 future calendar years as columns
}
\description{
Summarizes single claim cashflow predictions to origin year cashflows predictions
}
\examples{
cashflow <- matrix(0, nrow = 5, ncol = 250)
rownames(cashflow) <- paste0("Claim", as.character(1:5))
cashflow[c(1:10, 13:15, 18:20, 25)] <- 1000
claims_orig_years <- c(2010, 2010, 2010, 2012, 2012)
# printing only the first 5 of 250 columns
print(cbind(cashflow[,1:5], claims_orig_years))
convert_single2dev_year_prediction(cashflow, claims_orig_years, 2010, 2012)
}
