% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{convert_single2dev_year_triangle}
\alias{convert_single2dev_year_triangle}
\title{Generating triangle per development year from historic cashflow of single claims}
\usage{
convert_single2dev_year_triangle(
  cashflow,
  claims_orig_years,
  first_orig_year,
  last_orig_year
)
}
\arguments{
\item{cashflow}{matrix with one row per claim and columns first_orig_year:last_orig_year}

\item{claims_orig_years}{vector with origin years corresponding to claims in 'cashflow'}

\item{first_orig_year}{desired first origin year}

\item{last_orig_year}{desired last origin year}
}
\value{
matrix containing claims history triangle
}
\description{
Creates a matrix with each historic origin year as row and as calendar year in columns and sums in upper right triangle of the matrix
}
\details{
last_orig_year will automatically be treated as the last observed calendar year, so the resulting matrix is quadratic
}
\examples{
cashflow <- matrix(0, nrow = 5, ncol = 3)
rownames(cashflow) <- paste0("Claim", as.character(1:5))
colnames(cashflow) <- 2010:2012
cashflow[c(1, 2, 3, 6, 7, 12, 14, 15)] <- 1000
claims_orig_years <- c(2010, 2010, 2010, 2012, 2012)
print(cbind(cashflow, claims_orig_years))
convert_single2dev_year_triangle(cashflow, claims_orig_years, 2010, 2012)
}
