% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{cum2inc}
\alias{cum2inc}
\title{cumulative to incremental claims triangle}
\usage{
cum2inc(cum_triangle)
}
\arguments{
\item{cum_triangle}{numeric matrix}
}
\value{
numeric matrix
}
\description{
Transforms a cumulative triangle or matrix into an incremental triangle or matrix.
}
\examples{
cum2inc(matrix(c(1000, 800, 1200, 1200, 900, 1200, 1300, 900, 1200), 3))
}
