% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{cut2history}
\alias{cut2history}
\title{Cuts out the history part of a matrix with historic and future payments}
\usage{
cut2history(total_matrix)
}
\arguments{
\item{total_matrix}{numeric matrix with n rows and n + 250 columns. Rownames must be the origin years, colnames
must be the calendar years}
}
\value{
First n columns of total_matrix
}
\description{
Cuts out the history part of a matrix with historic and future payments
}
\examples{
total_matrix <- cbind(skeleton_triangle(2000,2010), skeleton_prediction(2000, 2010))
cut2history(total_matrix)
}
