% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{cut2prediction}
\alias{cut2prediction}
\title{Cuts out the prediction part of a matrix with historic and predicted payments}
\usage{
cut2prediction(total_matrix)
}
\arguments{
\item{total_matrix}{numeric matrix with n rows and n + 250 columns. Rownames must be the origin years, colnames
must be the calendar years}
}
\value{
Last 250 columns of total_matrix
}
\description{
Cuts out the prediction part of a matrix with historic and predicted payments
}
\examples{
total_matrix <- cbind(skeleton_triangle(2000,2010), skeleton_prediction(2000, 2010))
cut2prediction(total_matrix)
}
