% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_checks.R
\name{exposure_ok}
\alias{exposure_ok}
\title{Checking exposure dataframe for possible mistakes}
\usage{
exposure_ok(exposure, first_orig_year, last_orig_year)
}
\arguments{
\item{exposure}{Dataframe that must contain one row for each origin year between \code{first_orig_year}
and \code{last_orig_year} and the columns \code{Origin_year} and \code{Exposure}.}

\item{first_orig_year}{First origin year.}

\item{last_orig_year}{Last origin year.}
}
\value{
TRUE if every check is passed, else FALSE.
}
\description{
This function checks this input data for possible mistakes, implausible entries or missing or double rows.
}
\examples{
exposure_ok(exposure_xmpl, 1989, 2023)
}
