% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exposure_xmpl}
\alias{exposure_xmpl}
\title{Example of an exposure dataframe}
\format{
\subsection{\code{exposure_xmpl}}{

A data frame with 35 rows and 2 columns:
\describe{
\item{Origin_year}{Origin year}
\item{Exposure}{Numeric exposure, e.g. premiums or number of contracts}
}
}
}
\source{
https://www.gdv.de/gdv/statistik/statistiken-zur-deutschen-versicherungswirtschaft-uebersicht/schaden-und-unfallversicherung/anzahl-der-vertraege-und-schaeden-in-der-schaden-unfallversicherung-137946
}
\usage{
exposure_xmpl
}
\description{
Dataframe contains one row per origin year and one corresponding exposure
(here number of contracts of the german insurance market).
}
\keyword{datasets}
