% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_claims.R
\name{fill_up_triangle}
\alias{fill_up_triangle}
\title{Filling up claims triangle to rectangle}
\usage{
fill_up_triangle(triangle, pattern = NULL)
}
\arguments{
\item{triangle}{cumulative claims triangle. An (mxn)-matrix where only the upper left triangle (including diagonal) will be used}

\item{pattern}{Default: NULL, numeric vector or NULL if no external pattern shall be used
The pattern vector may be longer than the triangle for applying a tail pattern.}
}
\value{
filled triangle. numeric matrix
}
\description{
uses a claims development pattern to calculate the lower right triangle of a matrix
}
\examples{
fill_up_triangle(matrix(c(1000, 800, 1200, 1200, 900, 1200, 1300, 900, 1200), 3))
fill_up_triangle(matrix(c(1000, 800, 1200, 1200, 900, 1200, 1300, 900, 1200), 3),
                        c(1.2, 1.1, 1.05, 1.01))
}
