% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{filter_large_claims}
\alias{filter_large_claims}
\title{Filtering extended_claims_data to large claims and add derived columns}
\usage{
filter_large_claims(
  extended_claims_data,
  threshold,
  first_orig_year,
  expected_year_of_growing_large = 3
)
}
\arguments{
\item{extended_claims_data}{dataframe after step 4, see details of \code{\link[=prepare_data]{prepare_data()}}.}

\item{threshold}{numeric value to separate small from large claims, see details of \code{\link[=prepare_data]{prepare_data()}}.}

\item{first_orig_year}{Integer value for the first origin year for which a full claim history is available, see details of \code{\link[=prepare_data]{prepare_data()}}.}

\item{expected_year_of_growing_large}{Integer Value for claims without complete history, see details, see details of \code{\link[=prepare_data]{prepare_data()}}.}
}
\value{
dataframe filtered to large claims
}
\description{
The indexed columns allow for the exact calculation of \strong{which} claim has become large and \strong{when}.
Claims that have not become large yet are eliminated. \cr \cr
\strong{This function resembles only step 5 of the data preparation process, see description in details of \code{\link[=prepare_data]{prepare_data()}}}
}
\examples{
# this example uses data provided with this package
library(dplyr)
new_data <- minimal_claims_data_xmpl \%>\%
   reduce_data(indices_xmpl, 400000) \%>\%
   add_missing_years(2023) \%>\%
   add_columns() \%>\%
   add_indexed_columns(indices_xmpl) \%>\%
   filter_large_claims(400000, 1989, 3)
head(new_data)
}
