% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annuity_functions.R
\name{generate_annuities_prediction}
\alias{generate_annuities_prediction}
\title{Generate annuities prediction per development year from rolled out single annuities}
\usage{
generate_annuities_prediction(
  annuities,
  annuities_rolled_out,
  first_orig_year,
  last_orig_year
)
}
\arguments{
\item{annuities}{dataframe, see description of \code{pool_of_annuities} in details of \code{\link[=prepare_data]{prepare_data()}}.}

\item{annuities_rolled_out}{matrix as result of \code{\link[=roll_out_annuity]{roll_out_annuity()}}}

\item{first_orig_year}{Integer value for the first origin year, will be treated as the first observed calendar year.}

\item{last_orig_year}{Integer value for the last origin year which will also be treated as the last calendar year,
usually the calendar year that just ended.}
}
\value{
numeric matrix with one row per year between \code{first_orig_year} and \code{last_orig_year} and 250 columns for
predicted future annuity payments.
}
\description{
Generate annuities prediction per development year from rolled out single annuities
}
\examples{
# this example uses data provided with this package
print(minimal_active_annuities_xmpl)

# Create payments matrix
payments <- generate_annuity_payments(annuities = minimal_active_annuities_xmpl,
                                      last_orig_year = 2023)
print(payments[,1:10])

# Create probability matrix
probs <- generate_annuity_probabilities(annuities = minimal_active_annuities_xmpl,
                                        mortality = mortality_xmpl,
                                        age_shift = age_shift_xmpl)
print(probs[,1:10])

annuities_rolled_out <- roll_out_annuity(payments = payments,
                                         probs = probs,
                                         stop.at = "random_death_year")

# only print the first 10 columns
generate_annuities_prediction(annuities = minimal_active_annuities_xmpl,
                              annuities_rolled_out = annuities_rolled_out,
                              first_orig_year = 1989,
                              last_orig_year = 2023)[,1:10]
}
