% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annuity_functions.R
\name{generate_annuity_probabilities}
\alias{generate_annuity_probabilities}
\title{Create matrix of future mortality probabilities}
\usage{
generate_annuity_probabilities(annuities, mortality, age_shift)
}
\arguments{
\item{annuities}{dataframe, see description of \code{pool_of_annuities} in details of \code{\link[=prepare_data]{prepare_data()}}.}

\item{mortality}{dataframe, see description of \link{mortality_xmpl}.}

\item{age_shift}{dataframe, see description of \link{age_shift_xmpl}.}
}
\value{
dataframe with one row per row in 'annuities' and 130 columns for future mortality probabilities.
}
\description{
Creates a matrix with one row per annuitiy and mortality probabilities in columns.
}
\details{
In order to reduce computing time when rolling out annuities, they are split into two matrices, one with agreed
future payments and one with the corresponding survival probabilities. This is the function for the latter. \cr \cr
The function uses the biometric entries of the dataframe \code{annuities} to
create the matrix of the corresponding probabilities. \cr \cr
The first column is always 0 due to the assumption that annuities are paid in advance at the start of the year. \cr \cr
}
\examples{
# this example uses data provided with this package
print(minimal_active_annuities_xmpl)

# show only the first 10 columns
generate_annuity_probabilities(annuities = minimal_active_annuities_xmpl,
                               mortality = mortality_xmpl,
                               age_shift = age_shift_xmpl)[,1:10]
}
