% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{generate_claims_list}
\alias{generate_claims_list}
\title{Generate claims list}
\usage{
generate_claims_list(extended_claims_data, first_orig_year, last_orig_year)
}
\arguments{
\item{extended_claims_data}{prepared dataframe, see details of \code{\link[=prepare_data]{prepare_data()}}.}

\item{first_orig_year}{Integer value for the first origin year with complete history.}

\item{last_orig_year}{Integer value for the last origin year which will also be treated as the last calendar year,
usually the calendar year just ended.}
}
\value{
dataframe with current large claim status
}
\description{
The prepared data will be reduced to claims with origin year between \code{first_orig_year} and \code{last_orig_year}
and to the calendar year \code{last_orig_year}. Only necessary columns will be returned.
}
\examples{
# this example uses data provided with this package
extended_claims_data <- prepare_data(claims_data = minimal_claims_data_xmpl,
                                     indices = indices_xmpl,
                                     threshold = 400000,
                                     first_orig_year = 1989,
                                     last_orig_year = 2023,
                                     expected_year_of_growing_large = 3,
                                     reserve_classes = c(1, 200001, 400001, 700001, 1400001),
                                     pool_of_annuities = minimal_pool_of_annuities_xmpl)
claims_list <- generate_claims_list(extended_claims_data = extended_claims_data,
                                    first_orig_year = 1989,
                                    last_orig_year = 2023)
print(claims_list)

}
