% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{generate_history_per_claim}
\alias{generate_history_per_claim}
\title{Generating claims history from dataframe}
\usage{
generate_history_per_claim(data, column, first_orig_year, last_orig_year)
}
\arguments{
\item{data}{dataframe with at least the columns 'Claim_id', 'Calendar_year' and col_name}

\item{column}{column to sum across}

\item{first_orig_year}{desired first origin year, will be treated as the first observed calendar year}

\item{last_orig_year}{desired last origin year, will be treated as the last observed calendar year}
}
\value{
matrix containing claims history per claim
}
\description{
Creates a matrix with one row per claim and one column per calendar year
}
\examples{
claims_data <- data.frame(Claim_id = c(rep("Claim1", 3), "Claim2", rep("Claim3", 2)),
                          Origin_year = c(rep(2010, 4), rep(2012, 2)),
                          Calendar_year = c(2010:2012, 2012, 2012, 2014),
                          Dev_year_since_large = c(0, 1, 2, 1, 1, 2),
                          Cl_payment_cal = rep(1000, 6))
print(claims_data)
generate_history_per_claim(claims_data, "Cl_payment_cal", 2009, 2014)
}
