% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibnr.R
\name{generate_ibnr_pools}
\alias{generate_ibnr_pools}
\title{Generate ibnr pools}
\usage{
generate_ibnr_pools(
  large_claims_list,
  first_orig_year,
  last_orig_year,
  years_for_ibnr_pools
)
}
\arguments{
\item{large_claims_list}{Dataframe with one row per known large claim generated by \code{\link[=generate_claims_list]{generate_claims_list()}}.}

\item{first_orig_year}{First origin year with complete history.}

\item{last_orig_year}{Last origin year.}

\item{years_for_ibnr_pools}{Vector containing the calendar years that shall be used to build the pools.
Thus ibnr claims that have become large in other calendar years don't affect the model.}
}
\value{
List of one vector per development year with each vector containing the row numbers of the
claims in \code{large_claims_list} that have become large in that development year and in a calendar year
in \code{years_for_ibnr_pools}.
}
\description{
For sampling purposes this pool shall contain a list with one vector per development year.
The vector contains historic ibnr claims.
}
\examples{
# prepare data
extended_claims_data_xmpl <- prepare_data(claims_data = claims_data_xmpl,
                                          indices = indices_xmpl,
                                          threshold = 4e5,
                                          first_orig_year = 1989,
                                          last_orig_year = 2023,
                                          expected_year_of_growing_large = 3,
                                          reserve_classes = c(1, 200001, 400001, 700001, 1400001),
                                          pool_of_annuities = pool_of_annuities_xmpl)

# generate claims list
large_claims_list_xmpl <- generate_claims_list(extended_claims_data = extended_claims_data_xmpl,
                                               first_orig_year = 1989,
                                               last_orig_year = 2023)
generate_ibnr_pools(large_claims_list_xmpl, 1989, 2023, 2014:2023)
}
