% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_claims.R
\name{generate_pattern}
\alias{generate_pattern}
\title{Claims Development Pattern}
\usage{
generate_pattern(
  triangle,
  volume = NULL,
  external_pattern = NULL,
  int2ext_transition = NULL
)
}
\arguments{
\item{triangle}{cumulative claims triangle. An (mxn)-matrix where only the upper left triangle (including diagonal) will be used}

\item{volume}{Number of newest calendar years to be used. Default: NULL,
An Integer > 0 or NULL for no limitation}

\item{external_pattern}{numeric vector or NULL if no external pattern shall be used, Default: NULL}

\item{int2ext_transition}{year of transition from internal to external pattern. Integer or NULL for no transition.}
}
\value{
numeric vector
}
\description{
generating the claims development pattern from a claims triangle and mixing it with an external pattern, e.g. for considering tail patterns
}
\examples{
generate_pattern(matrix(c(1000, 800, 1200, 1200, 900, 1200, 1300, 900, 1200), 3))
generate_pattern(matrix(c(1000, 800, 1200, 1200, 900, 1200, 1300, 900, 1200), 3), 1,
                        c(1.2, 1.1, 1.05, 1.01), 3)
}
