% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pools.R
\name{generate_pools}
\alias{generate_pools}
\title{Generate pools to sample from}
\usage{
generate_pools(
  extended_claims_data,
  reserve_classes,
  years_for_pools,
  start_of_tail,
  end_of_tail,
  lower_outlier_limit = -Inf,
  upper_outlier_limit = Inf,
  pool_of_annuities
)
}
\arguments{
\item{extended_claims_data}{Dataframe as output of \code{\link[=prepare_data]{prepare_data()}}.}

\item{reserve_classes}{Numeric vector to specify reserve classes at the niveau of \code{index_year}, see details of \code{\link[=prepare_data]{prepare_data()}}.}

\item{years_for_pools}{Vector containing the calendar years that shall be used to build the pools.}

\item{start_of_tail}{First year to be treated as a tail year.}

\item{end_of_tail}{Last development year.}

\item{lower_outlier_limit}{Numeric value to specify lower outlier limit at the niveau of \code{index_year} see details of \code{\link[=prepare_data]{prepare_data()}}.
Lower payments will be set to lower_outlier_limit.}

\item{upper_outlier_limit}{Numeric value to specify upper outlier limit at the niveau of \code{index_year}, see details of \code{\link[=prepare_data]{prepare_data()}}.
Higher payments will be set to upper_outlier_limit.}

\item{pool_of_annuities}{Pool of annuities that was used to generate \code{extended_claims_data}.}
}
\value{
List of three objects, see details.
}
\description{
Generate pools to sample from
}
\details{
The function returns a list of three objects:
\enumerate{
\item \strong{pretail pools} \cr
This is a list containing one list per reserve class which contain one dataframe per development year. These dataframes
are the pools of historic claims from which to sample if a newer claim has the corresponding development year and reserve class.
It consists of the following columns:
\itemize{
\item \code{Ind_cl_payment_cal} is the indexed calendar year payment.
\item \code{Exit_reserve_class} is the "new" reserve class after the year.
\item \code{New_annuity_1} to \code{New_annuity_5} are the indices of new annuities in the dataframe \code{pool_of_annuities}.
\item \code{Ind_entry_cl_reserve} is the entry claim reserve which is only used for analysing purposes, for example to identify special claims.
}
\item \strong{tail pools} \cr
This is a list containing one dataframe per class. No distinction is made between development years except in the indexed
calendar year payments. These are raised to the niveau of the first tail year. The columns are the same as in 1.
\item \strong{pool of annuities} \cr
As the first two lists refer to the index in this pool, it has to be attached to enable sampling.
}

Note that in the pools list the first reserve class is always reserve class 0 and the first development year is always development year 2!
}
\examples{
# this example uses data provided with this package
extended_claims_data <- prepare_data(claims_data = minimal_claims_data_xmpl,
                                     indices = indices_xmpl,
                                     threshold = 400000,
                                     first_orig_year = 1989,
                                     last_orig_year = 2023,
                                     expected_year_of_growing_large = 3,
                                     reserve_classes = c(1, 200001, 400001, 700001, 1400001),
                                     pool_of_annuities = minimal_pool_of_annuities_xmpl)

pools <- generate_pools(extended_claims_data = extended_claims_data,
                        reserve_classes = c(1, 200001, 400001, 700001, 1400001),
                        years_for_pools = 2014:2023,
                        start_of_tail = 17,
                        end_of_tail = 50,
                        lower_outlier_limit = -Inf,
                        upper_outlier_limit = Inf,
                        pool_of_annuities = minimal_pool_of_annuities_xmpl)
# pretail pools reserve class 0 and development year 2
print(pools[[1]][[1]][[1]])

# tail pools reserve class 2
print(pools[[2]][[3]])

# pool of annuities
print(pools[[3]])
}
