% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{generate_triangle}
\alias{generate_triangle}
\title{Generating claims history triangle from dataframe}
\usage{
generate_triangle(
  data,
  col_name,
  first_orig_year,
  last_orig_year,
  include_small_claims = FALSE
)
}
\arguments{
\item{data}{dataframe with at least the columns 'Origin_year', 'Calendar_year' and col_name}

\item{col_name}{column to sum across}

\item{first_orig_year}{desired first origin year}

\item{last_orig_year}{desired last origin year}

\item{include_small_claims}{Default: FALSE. Set to FALSE to only include history as large claim. See details.}
}
\value{
matrix containing claims history triangle
}
\description{
Creates a matrix with each historic origin year as row and as calendar year in columns and sums of col_name entrys in upper right triangle of the matrix
}
\details{
last_orig_year will automatically be treated as the last observed calendar year, so the resulting matrix is quadratic /cr
if include_small_claims = FALSE, data must contain an additional column named "Dev_year_since_large to separate small from large claim history
}
\examples{
claims_data <- data.frame(Claim_id = c(rep("Claim1", 3), "Claim2", rep("Claim3", 2)),
                          Origin_year = c(rep(2010, 4), rep(2012, 2)),
                          Calendar_year = c(2010:2012, 2012, 2012, 2014),
                          Dev_year_since_large = c(0, 1, 2, 1, 1, 2),
                          Cl_payment_cal = rep(1000, 6))
print(claims_data)
generate_triangle(claims_data, "Cl_payment_cal", 2009, 2014)
}
