% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reinsurance.R
\name{generate_xl_indices_per_claim}
\alias{generate_xl_indices_per_claim}
\title{Helper function to create a matrix with indices per claim stemming from reinsurance treaty clause}
\usage{
generate_xl_indices_per_claim(
  claims_list,
  reinsurance,
  indices,
  first_orig_year,
  last_orig_year
)
}
\arguments{
\item{claims_list}{dataframe of claims generated with \code{\link[=generate_claims_list]{generate_claims_list()}}.}

\item{reinsurance}{dataframe of reinsurance structures, see details of \code{\link[=xl_cashflow]{xl_cashflow()}}.}

\item{indices}{dataframe for indexation, see details of \code{\link[=prepare_data]{prepare_data()}}.}

\item{first_orig_year}{integer value for the first origin year, will be treated as the first observed calendar year.}

\item{last_orig_year}{integer value for the last origin year which will also be treated as the last calendar year,
usually the calendar year that just ended.}
}
\value{
numeric matrix with one row per claim and one column for each past and 250 future calendar years. Entries
for years before the claim origin year are set to 1. The other entries are indexation factors.
}
\description{
This helper function converts the dataframes \code{reinsurance} and \code{indices} into a matrix with
one row per claim and one column for each past year from \code{first_orig_year} to \code{last_orig_year} and
for 250 future calendar years. \cr
This matrix will be used to properly calculate the reinsurance payments.
}
\examples{
# this example uses data provided with this package
# Create large claims list
extended_claims_data <- prepare_data(claims_data = minimal_claims_data_xmpl,
                                     indices = indices_xmpl,
                                     threshold = 400000,
                                     first_orig_year = 1989,
                                     last_orig_year = 2023,
                                     expected_year_of_growing_large = 3,
                                     reserve_classes = c(1, 200001, 400001, 700001, 1400001),
                                     pool_of_annuities = minimal_pool_of_annuities_xmpl)

large_claims_list <- generate_claims_list(extended_claims_data = extended_claims_data,
                                          first_orig_year = 1989,
                                          last_orig_year = 2023)

print(reinsurance_xmpl)

# show only first 20 columns
generate_xl_indices_per_claim(claims_list = large_claims_list,
                              reinsurance = reinsurance_xmpl,
                              indices = indices_xmpl,
                              first_orig_year = 1989,
                              last_orig_year = 2023)[,1:20]

}
