% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibnr.R
\name{get_additive_ibnr_model}
\alias{get_additive_ibnr_model}
\title{Generating additive model to estimate future ibnr claim numbers}
\usage{
get_additive_ibnr_model(
  large_claims_list,
  first_orig_year,
  last_orig_year,
  exposure,
  years_for_ibnr_pools
)
}
\arguments{
\item{large_claims_list}{Dataframe with one row per known large claim as a result of \code{\link[=generate_claims_list]{generate_claims_list()}}}

\item{first_orig_year}{Desired first origin year.}

\item{last_orig_year}{Desired last origin year.}

\item{exposure}{Dataframe that must contain one row for each origin year between \code{first_orig_year} and \code{last_orig_year}
and the columns \code{Origin_year} and \code{Exposure}.}

\item{years_for_ibnr_pools}{Vector containing the calendar years that shall be used to build the pools.
Thus ibnr claims that have became large in other calendar years don't affect the model.}
}
\value{
Dataframe with columns \code{Dev_year}, \code{Ibnr_numbers}, \code{Sum_of_exposures} and \code{Factor}.
}
\description{
This function calculates the number of ibnr claims per development year and divides it by the
exposure these ibnr claims have raised from.
}
\examples{
# prepare data
extended_claims_data_xmpl <- prepare_data(claims_data = claims_data_xmpl,
                                          indices = indices_xmpl,
                                          threshold = 4e5,
                                          first_orig_year = 1989,
                                          last_orig_year = 2023,
                                          expected_year_of_growing_large = 3,
                                          reserve_classes = c(1, 200001, 400001, 700001, 1400001),
                                          pool_of_annuities = pool_of_annuities_xmpl)

# generate claims list
large_claims_list_xmpl <- generate_claims_list(extended_claims_data = extended_claims_data_xmpl,
                                               first_orig_year = 1989,
                                               last_orig_year = 2023)

# generate additive ibnr model using the last ten calendar years
get_additive_ibnr_model(large_claims_list = large_claims_list_xmpl,
                        first_orig_year = 1989,
                        last_orig_year = 2023,
                        exposure = exposure_xmpl,
                        years_for_ibnr_pools = 2014:2023)

}
