% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibnr.R
\name{get_expected_ibnr_numbers}
\alias{get_expected_ibnr_numbers}
\title{Derive expected future ibnr numbers}
\usage{
get_expected_ibnr_numbers(additive_ibnr_model, exposure)
}
\arguments{
\item{additive_ibnr_model}{Dataframe as a result of \code{\link[=get_additive_ibnr_model]{get_additive_ibnr_model()}}.}

\item{exposure}{Dataframe that must contain one row for each desired origin year
and the columns \code{Origin_year} and \code{Exposure}.}
}
\value{
Expected ibnr numbers matrix with one row per origin year and one column per development year.
}
\description{
Expected future ibnr numbers are calculated from the additive ibnr model and the exposure.
}
\examples{
add_mod <- data.frame(Dev_year = 1:5,
                      Ibnr_numbers = c(100, 44, 15, 0, 2),
                      Sum_of_exposures = c(42000, 40000, 38000, 37000, 35000),
                      Factor = c(0.00238, 0.0011, 0.00039, 0, 0.00006))
exp <- data.frame(Origin_year = 2020:2023,
                  Exposure = c(1000, 1200, 1500, 1900))
get_expected_ibnr_numbers(add_mod, exp)
}
