% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_checks.R
\name{historic_indices_ok}
\alias{historic_indices_ok}
\title{Checking historic_indices dataframe for possible mistakes}
\usage{
historic_indices_ok(historic_indices, first_orig_year, last_orig_year)
}
\arguments{
\item{historic_indices}{Dataframe, see description of \code{\link[=expand_historic_indices]{expand_historic_indices()}}.}

\item{first_orig_year}{First origin year with full history.}

\item{last_orig_year}{Last origin year.}
}
\value{
TRUE if every check is passed, else FALSE.
}
\description{
This function checks this input data for possible mistakes, implausible entries or missing or double rows.
}
\examples{
historic_indices_ok(historic_indices_xmpl, 2015, 2023)
}
