% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{historic_indices_xmpl}
\alias{historic_indices_xmpl}
\title{Example of a historic_indices dataframe}
\format{
\subsection{\code{historic_indices_xmpl}}{

A data frame with 64 rows and 3 columns:
\describe{
\item{Calendar_year}{Calendar year}
\item{Index_gross}{is the claim payment development from one year to another, e.g. 0.02 for 2\% increase}
\item{Index_re}{is the contractually fixed claim payment development that is to be used in special index clauses that are a common part of longtail xl resinsurance programs.}}
}
}
\source{
\url{https://data.worldbank.org/indicator/FP.CPI.TOTL.ZG}
}
\usage{
historic_indices_xmpl
}
\description{
Dataframe contains one row per calendar year from 1960 to 2023 and corresponding inflation of consumer prices
in germany that may be used for indexating gross and reinsurance payments and reserves.
}
\keyword{datasets}
