% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibnr.R
\name{ibnr_transform2dataframe}
\alias{ibnr_transform2dataframe}
\title{Transform exptected ibnr matrix to dataframe}
\usage{
ibnr_transform2dataframe(expected_rounded_ibnr_numbers)
}
\arguments{
\item{expected_rounded_ibnr_numbers}{Matrix as a result of \code{\link[=get_expected_ibnr_numbers]{get_expected_ibnr_numbers()}} and \code{\link[=round_expected_ibnr_numbers]{round_expected_ibnr_numbers()}}. \cr
Rownames must be equal to origin years!}
}
\value{
Dataframe with one row per expected future ibnr large claim and columns \code{Origin_year},
\code{Dev_year_of_growing_large} and \code{Dev_year_since_large}.
\itemize{
\item \code{Origin_year} is the origin year of the expected future ibnr claim.
\item \code{Dev_year_of_growing_large} is the expected development year in which the expected ibnr claim
exceeds the threshold.
\item \code{Dev_year_since_large} is the "new" development year where counting starts in the calendar year
the claim exceeded the threshold. As these claims are expected to exceed the threshold in the future,
this column must be < 1.
}
}
\description{
The matrix with rounded expected numbers of ibnr large claims is transformed to a dataframe with
one row per expected claim.
}
\examples{
exp_ibnr <- matrix(c(0,0,0,2,0,0,1,1,0,0,0,0,0,0,0,1), 4)
rownames(exp_ibnr) <- 2020:2023
colnames(exp_ibnr) <- 2:5
ibnr_transform2dataframe(exp_ibnr)

}
