% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{inc2cum}
\alias{inc2cum}
\title{incremental to cumulative claims triangle}
\usage{
inc2cum(incr_triangle)
}
\arguments{
\item{incr_triangle}{numeric matrix}
}
\value{
numeric matrix
}
\description{
Transforms an incremental triangle or matrix into a cumulative triangle or matrix.
}
\examples{
inc2cum(matrix(c(1000, 800, 1200, 200, 100, 0, 100, 0, 0), 3))
}
