% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_checks.R
\name{indices_ok}
\alias{indices_ok}
\title{Checking indices dataframe for possible mistakes}
\usage{
indices_ok(indices, first_orig_year, last_orig_year)
}
\arguments{
\item{indices}{Dataframe for indexation, see details of \code{\link[=prepare_data]{prepare_data()}}.}

\item{first_orig_year}{First origin year.}

\item{last_orig_year}{Last origin year.}
}
\value{
TRUE if every check is passed, else FALSE.
}
\description{
This function checks this input data for possible mistakes, implausible entries or missing or double rows.
}
\examples{
indices_ok(indices_xmpl, 1989, 2023)
}
