% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indices_xmpl}
\alias{indices_xmpl}
\title{Example of a indices dataframe}
\format{
\subsection{\code{indices_xmpl}}{

A data frame with 286 rows and 4 columns:

\describe{
\item{Calendar_year}{Calendar year}
\item{Index_gross}{is the claim payment development from one year to another, e.g. 0.02 for 2\% increase}
\item{Index_re}{is the contractually fixed claim payment development that is to be used in special index clauses that are a common part of longtail xl resinsurance programs.}
\item{Transition_factor}{The transition_factor for year j indexes a payment of year j to the niveau of payments in the fixed index_year.}
}
}
}
\source{
{
Dataframe indices_xmpl can be generated by:

\if{html}{\out{<div class="sourceCode">}}\preformatted{library(dplyr)
indices_xmpl <- expand_historic_indices(historic_indices_xmpl,
                                        first_orig_year = 1989,
                                        last_orig_year = 2023,
                                        index_gross_future = 0.03,
                                        index_re_future = 0.02) \%>\%
                add_transition_factor(index_year = 2022)
}\if{html}{\out{</div>}}

}
}
\usage{
indices_xmpl
}
\description{
Dataframe contains one row for each historic calendar year and 250 future calendar years. \cr
This dataframe emerges from the dataframe \code{historic_indices_xmpl} and an expected future constant gross index of 3\%
and an expected future constant reinsurance index of 2\%.
}
\keyword{datasets}
