% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{matrix2triangle}
\alias{matrix2triangle}
\title{transforms a matrix into a lower right triangle}
\usage{
matrix2triangle(matr)
}
\arguments{
\item{matr}{numeric matrix}
}
\value{
numeric matrix
}
\description{
sets the upper left triangle including the diagonal to 0
}
\examples{
\dontrun{
matrix2triangle(matrix(c(1000, 800, 1200, 1200, 900, 1200, 1300, 900, 1250), 3))
}
}
\keyword{internal}
