% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_claims.R
\name{merge_results}
\alias{merge_results}
\title{merge paid and incurred results}
\usage{
merge_results(
  paid_result,
  incurred_result,
  weight_paid = 1,
  weight_incurred = 1
)
}
\arguments{
\item{paid_result}{ultimate vector or filled up cumulative paid claims triangle. numeric vector or numeric matrix}

\item{incurred_result}{ultimate vector or filled up cumulative incurred claims triangle. numeric vector or numeric matrix}

\item{weight_paid}{Default: 1. numeric vector of length 1 or same as number of rows of paid_result}

\item{weight_incurred}{Default: 1. numeric vector of length 1 or same as number of rows of incurred_result}
}
\value{
numeric vector
}
\description{
merging the results of paid and incurred triangle methods with given weights
}
\examples{
pai <- c(100, 200, 300)
filled_paid <- matrix(c(70, 150, 220, 90, 180, 270, 100, 200, 300), 3)
inc <- c(110, 190, 290)
merge_results(pai, inc, 1, 1)
merge_results(pai, inc, c(2,1,0), 1)
merge_results(filled_paid, inc, c(2,1,0), 1)

}
