% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_checks.R
\name{mortality_ok}
\alias{mortality_ok}
\title{Checking mortality dataframe for possible mistakes}
\usage{
mortality_ok(mortality)
}
\arguments{
\item{mortality}{Dataframe, see description of \link{mortality_xmpl}.}
}
\value{
TRUE if every check is passed, else FALSE.
}
\description{
This function checks this input data for possible mistakes, implausible entries or missing or double rows.
}
\examples{
mortality_ok(mortality_xmpl)
}
