% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mortality_xmpl}
\alias{mortality_xmpl}
\title{(unrealistic!) example of a mortality dataframe}
\format{
\subsection{\code{mortality_xmpl}}{

A data frame with 264 rows and 3 columns:
\describe{
\item{Gender}{"m" or "f" for male or female}
\item{Age}{Age between -10 and 121.}
\item{Probability}{mortality probabilities}
}
}
}
\source{
{Made up dataframe to show functionalities of the package.}
}
\usage{
mortality_xmpl
}
\description{
Dataframe contains one row for each combination of gender and age. \cr \cr
This data belongs to the set of \code{mortality_xmpl}, \code{age_shift_xmpl} and is totally made up to show functionalities of the package. \cr \cr
The age shift may lead to a negative technical age which must be considered in this dataframe.
}
\keyword{datasets}
