% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{nice_format}
\alias{nice_format}
\title{Easy to read formatting for numbers}
\usage{
nice_format(number, big.mark = ".", decimal.mark = ",")
}
\arguments{
\item{number}{numeric}

\item{big.mark}{"." or ",". Default: "."}

\item{decimal.mark}{"." or ",". Default: ","}
}
\value{
character
}
\description{
Easy to read formatting for numbers
}
\examples{
nice_format(pi*1e6)
}
