% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_pools_overview}
\alias{plot_pools_overview}
\title{Four simple plots to visualize the pools and open claims}
\usage{
plot_pools_overview(pools, large_claims_list)
}
\arguments{
\item{pools}{List as output of \code{\link[=generate_pools]{generate_pools()}}.}

\item{large_claims_list}{Dataframe with one row per known large claim generated by \code{\link[=generate_claims_list]{generate_claims_list()}}.}
}
\value{
List of four ggplot objects. \cr
\itemize{
\item plot_obs_numbers shows the observation numbers in the pools by reserve class and development year
\item plot_stack stacks the reserve classes
\item plot_reserves_allocation shows how the reserves of open claims are distributed to the reserve classes
\item plot_numbers_allocation shows how the numbers of open claims are distributed to the reserve classes
\item 250 columns for future calendar years
}
}
\description{
The function produces four simple ggplot objects to visualize the pools and
the distribution of open claims' reserves and open claims numbers to the
reserve classes.
}
\examples{
# this example uses data provided with this package
extended_claims_data <- prepare_data(claims_data = claims_data_xmpl,
                                     indices = indices_xmpl,
                                     threshold = 400000,
                                     first_orig_year = 1989,
                                     last_orig_year = 2023,
                                     expected_year_of_growing_large = 3,
                                     reserve_classes = c(1, 200001, 400001, 700001, 1400001),
                                     pool_of_annuities = pool_of_annuities_xmpl)

pools <- generate_pools(extended_claims_data = extended_claims_data,
                        reserve_classes = c(1, 200001, 400001, 700001, 1400001),
                        years_for_pools = 2014:2023,
                        start_of_tail = 17,
                        end_of_tail = 50,
                        lower_outlier_limit = -Inf,
                        upper_outlier_limit = Inf,
                        pool_of_annuities = pool_of_annuities_xmpl)

large_claims_list <- generate_claims_list(extended_claims_data = extended_claims_data,
                                          first_orig_year = 1989,
                                          last_orig_year = 2023)
p <- plot_pools_overview(pools = pools,
                         large_claims_list = large_claims_list)
}
