% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_checks.R
\name{pool_of_annuities_ok}
\alias{pool_of_annuities_ok}
\title{Checking pool_of_annuities dataframe for possible mistakes}
\usage{
pool_of_annuities_ok(pool_of_annuities)
}
\arguments{
\item{pool_of_annuities}{Dataframe of all historic annuities, see description of \code{pool_of_annuities} in
details of \code{\link[=prepare_data]{prepare_data()}}.}
}
\value{
TRUE if every check is passed, else FALSE.
}
\description{
This function checks this input data for possible mistakes, implausible entries or missing or double rows.
}
\examples{
pool_of_annuities_ok(pool_of_annuities_xmpl)
}
