% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pool_of_annuities_xmpl}
\alias{pool_of_annuities_xmpl}
\title{Example of a pool of annuities dataframe}
\format{
\subsection{\code{pool_of_annuities_xmpl}}{

A data frame with 208 rows and 11 columns:
\describe{
\item{Claim_id}{Claim ID}
\item{Annuity_id}{Annuity ID for the case of multiple annuities in one claim}
\item{Origin_year}{Origin year of the claim}
\item{Calendar_year}{Calendar year}
\item{Entering_year}{Year in which insurer and recipient have agreed on the annuity}
\item{Annuity_start}{Year in which payment starts, may be a past or a future year}
\item{Annuity_end}{Year in which payment ends, may be a past or a future year}
\item{Birth_year}{Birth year of recipient for assigning survival probability from mortality tables}
\item{Gender}{Gender of recipient for assigning mortality table}
\item{Annual_payment}{Agreed annual payment}
\item{Dynamic}{If a dynamic increase of the annual payment is part of the agreement, it can be specified here, e.g. 0.02 for 2\% annual payment increasement}
}
}
}
\source{
{A subset derived by reducing and anonymising an original dataset of an insurer.}
}
\usage{
pool_of_annuities_xmpl
}
\description{
Dataframe contains one row for each annuity that has ever been agreed on regardless
of whether the annuity is still active or not \cr
Data belongs to the set of \code{claims_data_xmpl}, \code{active_annuities_xmpl} and \code{pool_of_annuities_xmpl}.
}
\keyword{datasets}
