% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pools.R
\name{pools2dataframe}
\alias{pools2dataframe}
\title{Converts pools into a dataframe}
\usage{
pools2dataframe(pools)
}
\arguments{
\item{pools}{List as output of \code{\link[=generate_pools]{generate_pools()}}.}
}
\value{
Dataframe with the entries of pools plus the columns \code{Dev_year_since_large}
and \code{Entry_reserve_class}.
}
\description{
This function converts the first two of the three list entries of the object \code{pools} into an
easier to read dataframe.
}
\examples{
# this example uses data provided with this package
extended_claims_data <- prepare_data(claims_data = minimal_claims_data_xmpl,
                                     indices = indices_xmpl,
                                     threshold = 400000,
                                     first_orig_year = 1989,
                                     last_orig_year = 2023,
                                     expected_year_of_growing_large = 3,
                                     reserve_classes = c(1, 200001, 400001, 700001, 1400001),
                                     pool_of_annuities = minimal_pool_of_annuities_xmpl)

pools <- generate_pools(extended_claims_data = extended_claims_data,
                        reserve_classes = c(1, 200001, 400001, 700001, 1400001),
                        years_for_pools = 2014:2023,
                        start_of_tail = 17,
                        end_of_tail = 50,
                        lower_outlier_limit = -Inf,
                        upper_outlier_limit = Inf,
                        pool_of_annuities = minimal_pool_of_annuities_xmpl)

pools_as_df <- pools2dataframe(pools)
}
