% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare data for further usage}
\usage{
prepare_data(
  claims_data,
  indices,
  threshold,
  first_orig_year,
  last_orig_year,
  expected_year_of_growing_large = 3,
  reserve_classes,
  pool_of_annuities = NULL
)
}
\arguments{
\item{claims_data}{dataframe, see details.}

\item{indices}{dataframe, see details.}

\item{threshold}{numeric value, see details.}

\item{first_orig_year}{integer or numeric value, see details.}

\item{last_orig_year}{integer or numeric value, see details.}

\item{expected_year_of_growing_large}{integer or numeric value, see details.}

\item{reserve_classes}{numeric vector, see details.}

\item{pool_of_annuities}{dataframe, see details.}
}
\value{
dataframe of extended claims data filtered to large claims
}
\description{
The function filters large claims and adds columns that are needed for generating pools and for sampling. \cr
It summarises seven subfunctions, see details for detailed description of parameters and subfunctions.
}
\details{
Data preparation before applying \code{sicr} is done in seven steps. The subfunctions for each step can be used separately.
This allows users to skip steps in order to use own implementations instead. \cr \cr
\subsection{Description of needed input data}{
\itemize{
\item \code{first_orig_year} \cr
Integer value for the first origin year for which a full claim history is available. \cr
Information from older claims for which history is only available from \code{first_orig_year} to today may though be considered for the pools, see \code{claims_data}.
\item \code{last_orig_year} \cr
Integer value for the last origin year which will also be treated as the last calendar year, usually the calendar year just ended.
\item \code{claims_data} \cr
For each claim that could exceed the threshold after indexation this dataframe must contain one row
for every calendar year in which at least one of \code{Cl_reserve}, \code{An_reserve}, \code{Cl_payment_cal} or
\code{An_payment_cal} is not equal to 0. \cr \cr
To consider older claims for which history is only available from \code{first_orig_year} until now, add one row for this claim for calendar year \code{first_orig_year - 1}
and fill \code{Cl_payment_cal} and \code{An_payment_cal} with the cumulated payments up to that year and \code{Cl_reserve} and \code{An_reserve} with the
reserve at the end of \code{first_orig_year - 1}. This allows for checking if a claim belongs to large or small claims,
but due to the missing first part of the history the Dev_year_of_growing_large (see step 5) can't be determined for these claims. Dev_year_of_growing_large
has to be estimated instead (see parameter expected_year_of_growing_large). \cr \cr
claims_data must consist of the following columns:
\itemize{
\item \code{Claim_id} \cr type: character. Each claim must be assigned a unique claim id. These must correspond to the
claim id's in the dataframe \code{pool_of_annuities}.
\item \code{Origin_year} \cr type: integer or numeric. Origin year of the claim.
\item \code{Calendar_year} \cr type: integer or numeric. Calendar year of payments and reserves of the claim.
\item \code{Cl_payment_cal} \cr type: numeric. Claim payments made in this calendar year without annuity payments.
\item \code{Cl_reserve} \cr type: numeric. Claim reserve at the end of this calendar year without annuity reserves.
\item \code{An_payment_cal} \cr type: numeric. Annuity payments made in this calendar year.
\item \code{An_reserve} \cr type: numeric. Annuity reserve at the end of this calendar year.
}
\item \code{pool_of_annuities} \cr
This dataframe contains one row for each annuity that has ever been agreed on regardless of whether the annuity is still active or not. \cr \cr
\code{pool_of_annuities} must consist of the following columns:
\itemize{
\item \code{Claim_id} \cr type: character. Each claim must be assigned a unique claim id. These must correspond to the
claim id's in the dataframe \code{claims_data}.
\item \code{Annuity_id} \cr type: character. Annuity ID for the case of multiple annuities in one claim.
\item \code{Origin_year} \cr type: integer or numeric. Origin year of the claim.
\item \code{Calendar_year} \cr type: integer or numeric. Calendar year of information.
\item \code{Entering_year} \cr type: integer or numeric. Year in which insurer and recipient have agreed on the annuity.
\item \code{Annuity_start} \cr type: integer or numeric. Year in which payment starts, may be a past or a future year.
\item \code{Annuity_end} \cr type: integer or numeric. Year in which payment ends, may be a past or a future year.
\item \code{Birth_year} \cr type: integer or numeric. Birth year of recipient for assigning survival probability from mortality tables.
\item \code{Gender} \cr type: character "m" or "f". Gender of recipient for assigning mortality table.
\item \code{Annual_payment} \cr type: numeric. Agreed annual payment.
\item \code{Dynamic} \cr type: numeric. If a dynamic increase of the annual payment is part of the agreement, it can be specified here, e.g. 0.02 for 2\% annual payment increasement.
}
\item \code{indices} \cr
This dataframe contains one row for each historic calendar year and 250 future calendar years. \cr
Although the indices for the future calendar years (as well as the column \code{Index_re}) are not needed here but in the subsequent simulation,
it is required to create all rows and all columns of this dataframe just once with the helper functions.
\code{\link[=expand_historic_indices]{expand_historic_indices()}} and \code{\link[=add_transition_factor]{add_transition_factor()}}. \cr \cr
\code{indices} must consist of the following columns:
\itemize{
\item \code{Calendar_year} \cr type: integer or numeric.
\item \code{Index_gross} \cr type: numeric. Claim payment development from one year to another, e.g. 0.02 for 2\% increase
\item \code{Index_re} \cr type: numeric. Contractually fixed claim payment development that is to be used in special index
clauses that are a common part of longtail xl resinsurance programs.
\item \code{Transition_factor} \cr type: numeric. The transition_factor for year j indexes a payment of year j to the niveau
of payments in the fixed index_year.
}
\item \code{threshold} \cr
Numeric value to separate small from large claims. \code{Threshold} has to be stated at the niveau of index_year.
\item \code{reserve_classes} \cr numeric vector to specify reserve classes at the niveau of index_year, for example
\code{c(1, 1001)} for the three reserve classes \verb{(-Inf, 1)}, \verb{[1, 1001)} and \verb{[1001, Inf)}.
\item \code{expected_year_of_growing_large} \cr
Integer value for the estimated development year of becoming large for claims without full history (see description above). Default: 3.
}
}

\subsection{Step 1 - Reducing data}{

\code{claims_data} is reduced to possible large claims.
}

\subsection{Step 2 - Add missing rows}{

\code{claims_data} is requested to only contain rows in which one of \code{Cl_reserve}, \code{An_reserve}, \code{Cl_payment_cal} or \code{An_payment_cal} is not equal to 0. \cr
This step adds rows so that claims_data contains one row per calendar year between the origin year and the last origin year for each claim.
}

\subsection{Step 3 - Add columns}{

This step adds derived columns to \code{claims_data} that are needed for further operations:
\itemize{
\item \code{Development_year}
\item Sum of claim and annuity payments and reserve
\itemize{
\item \code{Payment_cal} as the sum of \code{Cl_payment_cal} and \code{An_payment_cal}
\item \code{Reserve} as the sum of \code{Cl_reserve} and \code{An_reserve}
}
\item cumulated columns
\itemize{
\item \code{Cl_payment_cum} as the cumulated claim payments
\item \code{An_payment_cum} as the cumulated annuity payments
\item \code{Payment_cum} as the sum of the two latter
}
\item incurred columns
\itemize{
\item \code{Cl_incurred} as the sum of \code{Cl_payment_cum} and \code{Cl_reserve}
\item \code{An_incurred} as the sum of \code{An_payment_cum} and \code{An_reserve}
\item \code{Incurred} as the sum of the two latter
}
\item Entry reserve columns
\itemize{
\item \code{Entry_cl_reserve} as the \code{Cl_reserve} of the beginning of the year
\item \code{Entry_an_reserve} as the \code{An_reserve} of the beginning of the year
\item \code{Entry_reserve} as the sum of the two latter
}
}
}

\subsection{Step 4 - Add indexed columns}{

This step uses the dataframe indices to index every payment and every reserve to the indexation year. \cr \cr
The following columns are calculated:
\itemize{
\item indexed payments
\itemize{
\item \code{Ind_cl_payment_cal}
\item \code{Ind_an_payment_cal}
\item \code{Ind_payment_cal} as the sum of the two latter
}
\item indexed reserves
\itemize{
\item \code{Ind_cl_reserve}
\item \code{Ind_an_reserve}
\item \code{Ind_reserve} as the sum of the two latter
}
\item indexed entry reserves
\itemize{
\item \code{Ind_entry_cl_reserve}
\item \code{Ind_entry_an_reserve}
\item \code{Ind_entry_reserve} as the sum of the two latter
}
\item cumulated columns
\itemize{
\item \code{Ind_cl_payment_cum}
\item \code{Ind_an_payment_cum}
\item \code{Ind_payment_cum} as the sum of the two latter
\item \code{Ind_cl_incurred}
\item \code{Ind_an_incurred}
\item \code{Ind_incurred} as the sum of the two latter
}
}
}

\subsection{Step 5 - Filtering large claims}{

The indexed columns allow for the exact calculation of \strong{which} claim has become large and \strong{when}.
Claims that have not become large yet are eliminated. \cr \cr
The following new columns are derived from this information:
\itemize{
\item \code{Large_since} as the calendar year in which the claim exceeded the threshold for the first time
\item \code{Dev_year_of_growing_large} as the development year in which the claim exceeded the threshold for the first time
\item \code{Dev_year_since_large} as the "new" development year where counting starts in year \code{Large_since} instead of the origin year
}
}

\subsection{Step 6 - Add reserve classes}{

The columns \code{Entry_reserve_class} and \code{Exit_reserve_class} can now be derived from \code{Ind_entry_cl_reserve} and \code{Ind_cl_reserve}.
}

\subsection{Step 7 - Attach future annuities}{

The columns \code{New_annuity_1} to \code{New_annuity_5} are added to the dataframe. If new annuities have been agreed on
in the calendar year and the claim of the row, these columns contain the row number of these annuities in the corresponding
dataframe \code{pool_of_annuities}. These columns are needed for the pool as the sampling shall consider claim payments and
new reserve classes as well as new annuities. \cr \cr

The function always adds \strong{five} columns for new annuities which is sufficient for the rare case that five new annuities are agreed on
for one single claim in one calendar year. Each further new annuity in one year will be ignored, which could lead to a slight underestimation
of the best estimate. To avoid this, data may be manually adjusted, for example by moving further annuities to another calendar year or
by aggregating annuities.
}
}
\examples{
# this example uses data provided with this package
new_data <- prepare_data(claims_data = minimal_claims_data_xmpl,
                         indices = indices_xmpl,
                         threshold = 400000,
                         first_orig_year = 1989,
                         last_orig_year = 2023,
                         expected_year_of_growing_large = 3,
                         reserve_classes = c(1, 200001, 400001, 700001, 1400001),
                         pool_of_annuities = minimal_pool_of_annuities_xmpl)
head(new_data)

}
