% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibnr.R
\name{prepare_ibnr}
\alias{prepare_ibnr}
\title{Prepares necessary ibnr objects}
\usage{
prepare_ibnr(
  large_claims_list,
  first_orig_year,
  last_orig_year,
  exposure = NULL,
  years_for_ibnr_pools = NULL
)
}
\arguments{
\item{large_claims_list}{Dataframe with one row per known large claim as a result of \code{\link[=generate_claims_list]{generate_claims_list()}}}

\item{first_orig_year}{Desired first origin year.}

\item{last_orig_year}{Desired last origin year.}

\item{exposure}{Dataframe that must contain one row for each origin year between \code{first_orig_year} and \code{last_orig_year}
and the columns \code{Origin_year} and \code{Exposure}. \cr
Default: NULL. In this case, a constant exposure is assumed.}

\item{years_for_ibnr_pools}{Vector containing the calendar years that shall be used to build the pools. \cr
Default: NULL. In this case, all years are used for building ibnr pools.}
}
\value{
List of two objects: \cr
\enumerate{
\item Dataframe with one row per expected future ibnr large claim and columns \code{Origin_year},
\code{Dev_year_of_growing_large} and \code{Dev_year_since_large}.
\itemize{
\item \code{Origin_year} is the origin year of the expected future ibnr claim.
\item \code{Dev_year_of_growing_large} is the expected development year in which the expected ibnr claim
exceeds the threshold.
\item \code{Dev_year_since_large} is the "new" development year where counting starts in the calendar year
the claim exceeded the threshold. As these claims are expected to exceed the threshold in the future,
this column must be < 1.
}
\item List of one vector per development year with each vector containing the row numbers of the
claims in \code{large_claims_list} that have become large in that development year and in a calendar year
in \code{years_for_ibnr_pools}.
}
}
\description{
The function generates the expected future ibnr claims and the pools of historic ibnr claims
from which the claim payments of the future ibnr claims will be sampled.
}
\details{
The first object is generated in four steps with the functions \code{\link[=get_additive_ibnr_model]{get_additive_ibnr_model()}}, \code{\link[=get_expected_ibnr_numbers]{get_expected_ibnr_numbers()}},
\code{\link[=round_expected_ibnr_numbers]{round_expected_ibnr_numbers()}} and \code{\link[=ibnr_transform2dataframe]{ibnr_transform2dataframe()}}. If one are more functions shall be
replaced by own implementations, the functions may be used separately.
}
\examples{
# prepare data
extended_claims_data_xmpl <- prepare_data(claims_data = claims_data_xmpl,
                                          indices = indices_xmpl,
                                          threshold = 4e5,
                                          first_orig_year = 1989,
                                          last_orig_year = 2023,
                                          expected_year_of_growing_large = 3,
                                          reserve_classes = c(1, 200001, 400001, 700001, 1400001),
                                          pool_of_annuities = pool_of_annuities_xmpl)

# generate claims list
large_claims_list_xmpl <- generate_claims_list(extended_claims_data = extended_claims_data_xmpl,
                                               first_orig_year = 1989,
                                               last_orig_year = 2023)

# generate example exposure with 5\% annual exposure increase since 1989

prepare_ibnr(large_claims_list_xmpl,
             1989,
             2023,
             exposure_xmpl,
             2014:2023)

}
