% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{reduce_data}
\alias{reduce_data}
\title{Reduce claims data to possible large claims}
\usage{
reduce_data(claims_data, indices, threshold)
}
\arguments{
\item{claims_data}{dataframe, see details of \code{\link[=prepare_data]{prepare_data()}}.}

\item{indices}{dataframe, see details of \code{\link[=prepare_data]{prepare_data()}}.}

\item{threshold}{numeric value to separate small from large claims, see details of \code{\link[=prepare_data]{prepare_data()}}.}
}
\value{
subset of dataframe \code{claims_data} where preselected small claims have been eliminated
}
\description{
The original dataframe is reduced to possible large claims to save computing time in the next steps. \cr \cr
This is not yet a precise separation as the indexed columns are missing at this point. \cr \cr
\strong{This function resembles only step 1 of the data preparation process, see description in details of \code{\link[=prepare_data]{prepare_data()}}}
}
\examples{
# this example uses data provided with this package
reduce_data(minimal_claims_data_xmpl, indices_xmpl, threshold = 400000)

}
