% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reinsurance_xmpl}
\alias{reinsurance_xmpl}
\title{Example of a reinsurance structures dataframe}
\format{
\subsection{\code{reinsurance_xmpl}}{

A data frame with 35 rows and 7 columns:
\describe{
\item{Origin_year}{Origin year}
\item{Priority}{Start of reinsurance coverage}
\item{Limit}{Maximum reinsurance coverage}
\item{Base_year}{The year from which the indexation shall start for this origin year}
\item{Margin_type}{One of the standard margin types \code{FIC}, \code{APK} or \code{SIC} or \code{none} for no reinsurance}
\item{Margin}{\code{APK} and \code{SIC} require a margin}
\item{Quote_share}{Quota_share to consider before applying the xl construction}
}
}
}
\source{
{Made up to show package functionalities.}
}
\usage{
reinsurance_xmpl
}
\description{
Dataframe contains one row per origin year and describes the reinsurance structure. \cr
Note: Only quota shares and XLs can be applied and no additional structures as AADs (aggregate annual deductable)
or AALs (aggregate annual limits). More details can be found in the details section of \code{\link[=xl_cashflow]{xl_cashflow()}}.
}
\keyword{datasets}
