% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_claims.R
\name{result2best_estimate}
\alias{result2best_estimate}
\title{deriving best estimate}
\usage{
result2best_estimate(result, paid)
}
\arguments{
\item{result}{ultimate vector or filled up cumulative claims triangle. numeric vector or numeric matrix}

\item{paid}{cumulative claims triangle. An (mxn)-matrix where only the upper left triangle (including diagonal) will be used}
}
\value{
best estimate. numeric vector
}
\description{
deriving best estimate from ultimate vector and paid triangle
}
\examples{
result2best_estimate(c(1300, 975, 1516.667),
                     matrix(c(1000, 800, 1200, 1200, 900, 1200, 1300, 900, 1200), 3))
result2best_estimate(matrix(c(1000, 800, 1200, 1200, 900, 1400, 1300, 975, 1516.667), 3),
                     matrix(c(1000, 800, 1200, 1200, 900, 1200, 1300, 900, 1200), 3))
}
