% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annuity_functions.R
\name{roll_out_active_annuities}
\alias{roll_out_active_annuities}
\title{Rolling out active annuities corresponding to large claims}
\usage{
roll_out_active_annuities(
  payments_active_annuities,
  probs_active_annuities,
  active_annuities,
  large_claims_list
)
}
\arguments{
\item{payments_active_annuities}{matrix of agreed future payments corresponding to active annuities, see \code{\link[=generate_annuity_payments]{generate_annuity_payments()}}.}

\item{probs_active_annuities}{matrix of mortality probabilities corresponding to active annuities, see \code{\link[=generate_annuity_probabilities]{generate_annuity_probabilities()}}.}

\item{active_annuities}{Dataframe of active annuities, see \link{active_annuities_xmpl}.}

\item{large_claims_list}{Dataframe of large claims generated with \code{\link[=generate_claims_list]{generate_claims_list()}}.}
}
\value{
Numeric matrix with one row per large claim and 130 future annuity payments.
}
\description{
Rolling out active annuities corresponding to large claims
}
\examples{
# this example uses data provided with this package
print(minimal_active_annuities_xmpl)

# Create payments matrix
payments <- generate_annuity_payments(annuities = minimal_active_annuities_xmpl,
                                      last_orig_year = 2023)
print(payments[,1:10])

# Create probability matrix
probs <- generate_annuity_probabilities(annuities = minimal_active_annuities_xmpl,
                                        mortality = mortality_xmpl,
                                        age_shift = age_shift_xmpl)
print(probs[,1:10])

# Create large claims list
extended_claims_data <- prepare_data(claims_data = minimal_claims_data_xmpl,
                                     indices = indices_xmpl,
                                     threshold = 400000,
                                     first_orig_year = 1989,
                                     last_orig_year = 2023,
                                     expected_year_of_growing_large = 3,
                                     reserve_classes = c(1, 200001, 400001, 700001, 1400001),
                                     pool_of_annuities = minimal_pool_of_annuities_xmpl)

large_claims_list <- generate_claims_list(extended_claims_data = extended_claims_data,
                                          first_orig_year = 1989,
                                          last_orig_year = 2023)

roll_out_active_annuities(payments_active_annuities = payments,
                          probs_active_annuities = probs,
                          active_annuities = minimal_active_annuities_xmpl,
                          large_claims_list = large_claims_list)
}
