% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annuity_functions.R
\name{roll_out_annuity}
\alias{roll_out_annuity}
\title{Rolling out one or more annuities}
\usage{
roll_out_annuity(payments, probs, stop.at = "random_death_year")
}
\arguments{
\item{payments}{numeric vector or matrix of agreed future payments with one row per annuity and 130 columns for future payments.}

\item{probs}{numeric vector or matrix with to \code{payments} corresponding mortality probabilities. Must be of same size as \code{payments}.}

\item{stop.at}{Chose between \code{random_death_year} (stochastic approach) and \code{end_of_probs} (deterministic approach).
The first samples a random death year of the annuity recipient and returns the agreed payments up to that year and 0 after.
The latter multiplies the payments with the corresponding mortality probabilities and returns a best estimate payment per year.}
}
\value{
numeric matrix of the same size as \code{payments} with future annuity payments.
}
\description{
Rolling out one or more annuities
}
\examples{
# this example uses data provided with this package
print(minimal_active_annuities_xmpl)

# Create payments matrix
payments <- generate_annuity_payments(annuities = minimal_active_annuities_xmpl,
                                      last_orig_year = 2023)
print(payments[,1:10])

# Create probability matrix
probs <- generate_annuity_probabilities(annuities = minimal_active_annuities_xmpl,
                                        mortality = mortality_xmpl,
                                        age_shift = age_shift_xmpl)
print(probs[,1:10])

# stochastic with random death year
roll_out_annuity(payments, probs, stop.at = "random_death_year")

# deterministic multiplying with mortality probabilities
roll_out_annuity(payments, probs, stop.at = "end_of_probs")
}
