% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special_claims.R
\name{roll_out_single_claim}
\alias{roll_out_single_claim}
\title{Converting a reserve into a cashflow by given patterns}
\usage{
roll_out_single_claim(
  reserve,
  years = 1,
  type = "linear",
  external_pattern = NULL
)
}
\arguments{
\item{reserve}{Numeric value of the reserve to be rolled out.}

\item{years}{Integer value stating how many future years shall be considered. Default: 1 \cr
This parameter is not used if \code{type} is set to \code{external}.}

\item{type}{Desired way of rolling out reserve. Must be one of \code{linear}, \code{constant} or \code{external}. Default: \code{linear}.\cr
In case of \code{linear} or \code{constant} the parameter \code{years} will be used. \code{external} requires the parameter \code{external_pattern}.}

\item{external_pattern}{Numeric vector with future payments as a percentage of \code{reserve}. Sum must be 1.}
}
\value{
Numeric vector of length 250 with future payments.
}
\description{
This is a helper function for the function \code{\link[=compute_special_claims]{compute_special_claims()}}. It is designed for the purpose of
rolling out a reserve of a special claim either by a given external pattern or linear or constant.
}
\examples{
# Default type linear and years 1
roll_out_single_claim(1000)
roll_out_single_claim(1000, 5)

# constant
roll_out_single_claim(1000, 5, "constant")

# external_pattern
roll_out_single_claim(1000, type = "external", external_pattern = c(0.3, 0.5, 0.1, 0.1, 0, 0))

}
