% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_claims.R
\name{rollout_best_estimate}
\alias{rollout_best_estimate}
\title{roll out best estimate to a cashflow}
\usage{
rollout_best_estimate(best_estimate, pattern, dev_year)
}
\arguments{
\item{best_estimate}{numeric vector.}

\item{pattern}{numeric vector of development factors per development year}

\item{dev_year}{integer vector > 0}
}
\value{
numeric vector of length 250
}
\description{
roll out best estimate to a cashflow
}
\note{
best_estimate and dev_year must have same length
}
\examples{
\dontrun{
rollout_best_estimate(1000, c(1.7, 1.3, 1.1, 1.05, 1, 1), 3)
}
}
\keyword{internal}
