% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibnr.R
\name{round_expected_ibnr_numbers}
\alias{round_expected_ibnr_numbers}
\title{Rounding expected ibnr matrix}
\usage{
round_expected_ibnr_numbers(expected_ibnr_numbers)
}
\arguments{
\item{expected_ibnr_numbers}{Matrix as a result of \code{\link[=get_expected_ibnr_numbers]{get_expected_ibnr_numbers()}}}
}
\value{
Matrix with same dimensions.
}
\description{
As only whole ibnr numbers can be used in the simulation, the expected ibnr matrix must be rounded. \cr \cr
Algorithms for rounding should allocate the rounded remainders to origin and development years
in a way that rowsums and colsums are barely changed. They must moreover secure that 0-entries remain 0 to
prevent subsequent errors. \cr \cr

This algorithm is one possible ways to do so and works very well in most examples. \cr \cr
\strong{Thanks to Michele Pellino for idea and implementation!}
}
\examples{
# small example
exp_ibnr <- matrix(c(0,0,0,2.09,0,0,0.585,0.741,0,0,0,0,0.06,0.072,0.09,0.114), 4)
rownames(exp_ibnr) <- 2020:2023
colnames(exp_ibnr) <- 2:5
# Unrounded:
print(exp_ibnr)
# Rounded:
round_expected_ibnr_numbers(exp_ibnr)

# more complex example
# prepare data
extended_claims_data_xmpl <- prepare_data(claims_data = claims_data_xmpl,
                                          indices = indices_xmpl,
                                          threshold = 4e5,
                                          first_orig_year = 1989,
                                          last_orig_year = 2023,
                                          expected_year_of_growing_large = 3,
                                          reserve_classes = c(1, 200001, 400001, 700001, 1400001),
                                          pool_of_annuities = pool_of_annuities_xmpl)

# generate claims list
large_claims_list_xmpl <- generate_claims_list(extended_claims_data = extended_claims_data_xmpl,
                                               first_orig_year = 1989,
                                               last_orig_year = 2023)

# generate additive ibnr model using the last ten calendar years
am <- get_additive_ibnr_model(large_claims_list = large_claims_list_xmpl,
                              first_orig_year = 1989,
                              last_orig_year = 2023,
                              exposure = exposure_xmpl,
                              years_for_ibnr_pools = 2014:2023)

expected_ibnr_numbers <- get_expected_ibnr_numbers(am, exposure_xmpl)
rounded_expected_ibnr_numbers <- round_expected_ibnr_numbers(expected_ibnr_numbers)
# Unrounded:
print(expected_ibnr_numbers)
# Rounded:
print(rounded_expected_ibnr_numbers)
# colsums before and after
print(data.frame(ColSums_unrounded = colSums(expected_ibnr_numbers),
                 ColSums_round = colSums(rounded_expected_ibnr_numbers)))

# rowsums before and after
print(data.frame(RowSums_unrounded = rowSums(expected_ibnr_numbers),
                 RowSums_round = rowSums(rounded_expected_ibnr_numbers)))

}
