% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{sample_1}
\alias{sample_1}
\title{Extention to Sample function}
\usage{
sample_1(pool, i)
}
\arguments{
\item{pool}{vector or list to choose from}

\item{i}{number of elements to choose}
}
\value{
vector of length i with elements drawn from pool
}
\description{
Out of a list x with only one element n, sample(x, m) returns m random numbers between 1 and n.
This function returns rep(n, m) instead in this case and behaves like the sample-function in any other case.
}
\examples{
\dontrun{
sample_1(1:10, 3)
sample_1(5, 2)
}
}
\keyword{internal}
