% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_claims.R
\name{settlement_rates}
\alias{settlement_rates}
\title{change development factors to settlement rates}
\usage{
settlement_rates(pattern)
}
\arguments{
\item{pattern}{numeric vector of development factors per development year}
}
\value{
numeric vector
}
\description{
change development factors to settlement rates
}
\examples{
\dontrun{
settlement_rates(c(1.7, 1.3, 1.1, 1, 1, 1))
}
}
\keyword{internal}
