% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sicr-package.R
\docType{package}
\name{sicr-package}
\alias{sicr-package}
\title{sicr: Stochastic Individual Claims Reserving}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dskowasch.github.io/sicr}
}

}
\author{
\strong{Maintainer}: Dirk Skowasch \email{d.skowasch@gmx.de} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
