\name{SICR}
\alias{SICR}
\title{Create a model for the SimInf framework}
\usage{
SICR(u0, tspan)
}
\arguments{
\item{u0}{A \code{data.frame} with the initial state in each node.}

\item{tspan}{A vector (length >= 2) of increasing time points
where the state of each node is to be returned.}
}
\description{
Create a model to be used by the SimInf framework.
}
\examples{
## Please add example(s) how to use the model
}
