% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{sicr_series}
\alias{sicr_series}
\title{Creating series of simulation results}
\usage{
sicr_series(
  n,
  columns = c("Claim_payments"),
  drop_ibnr_claims = TRUE,
  large_claims_list,
  first_orig_year,
  last_orig_year,
  pools,
  indices,
  history,
  exposure = NULL,
  years_for_ibnr_pools = NULL,
  active_annuities = NULL,
  age_shift = NULL,
  mortality = NULL,
  reinsurance = NULL,
  seed = 12345,
  progress = TRUE
)
}
\arguments{
\item{n}{Integer. Desired number of realisations.}

\item{columns}{Character vector containing the names of the desired information.
Valid names are the names of the matrices in the output of \code{\link[=sicr]{sicr()}}, see Value.
Default: c("Claim_payments")}

\item{drop_ibnr_claims}{TRUE if only known claims shall appear in the output. Default: TRUE}

\item{large_claims_list}{Dataframe with one row per known large claim generated by \code{\link[=generate_claims_list]{generate_claims_list()}}.}

\item{first_orig_year}{First origin year with complete history.}

\item{last_orig_year}{Last origin year.}

\item{pools}{List as output of \code{\link[=generate_pools]{generate_pools()}}.}

\item{indices}{Dataframe for indexation, see details of \code{\link[=prepare_data]{prepare_data()}}.}

\item{history}{numeric matrix with one row per claim with historic payments,
generated by \code{\link[=generate_history_per_claim]{generate_history_per_claim()}}. \cr
If generated in other ways, note that colnames must be the origin years. \cr
This matrix is needed to calculate ibnr payments and reinsurance if desired.}

\item{exposure}{Dataframe that must contain one row for each origin year between \code{first_orig_year}
and \code{last_orig_year} and the columns \code{Origin_year} and \code{Exposure}. \cr
Default: NULL. In this case, a constant exposure is assumed.}

\item{years_for_ibnr_pools}{Vector containing the calendar years that shall be used to build the pools. \cr
Default: NULL. In this case, all years are used for building ibnr pools.}

\item{active_annuities}{Dataframe of active annuities, see \link{active_annuities_xmpl}. Default: NULL.\cr
Only necessary if annuities shall be considered.}

\item{age_shift}{Dataframe, see description of \link{age_shift_xmpl}. Default: NULL \cr
Only necessary if annuities shall be considered.}

\item{mortality}{Dataframe, see description of \link{mortality_xmpl}. Default: NULL \cr
Only necessary if annuities shall be considered.}

\item{reinsurance}{Dataframe with reinsurance information, see details of \code{\link[=xl_cashflow]{xl_cashflow()}} or \link{reinsurance_xmpl}. Default: NULL\cr
Only necessary if reinsurance shall be considered.}

\item{seed}{Default: 12345. Setting a seed makes the simulation reproducible.}

\item{progress}{Default: TRUE. Set to FALSE if no progress bar shall be shown during simulation.}
}
\value{
Numeric three-dimensional array with \cr
\itemize{
\item one row per known claim followed by one row per expected ibnr claim if desired
\item 250 columns for future calendar years
\item one matrix for each of the n desired realisations
The entries contain the sum of the desired information stated in \code{columns}.
}
}
\description{
This function executes sicr n times and returns a result series
of each realisation for the desired columns.
}
\examples{
# this example uses data provided with this package
extended_claims_data <- prepare_data(claims_data = claims_data_xmpl,
                                     indices = indices_xmpl,
                                     threshold = 400000,
                                     first_orig_year = 1989,
                                     last_orig_year = 2023,
                                     expected_year_of_growing_large = 3,
                                     reserve_classes = c(1, 200001, 400001, 700001, 1400001),
                                     pool_of_annuities = pool_of_annuities_xmpl)

pools <- generate_pools(extended_claims_data = extended_claims_data,
                        reserve_classes = c(1, 200001, 400001, 700001, 1400001),
                        years_for_pools = 2014:2023,
                        start_of_tail = 17,
                        end_of_tail = 50,
                        lower_outlier_limit = -Inf,
                        upper_outlier_limit = Inf,
                        pool_of_annuities = pool_of_annuities_xmpl)

large_claims_list <- generate_claims_list(extended_claims_data = extended_claims_data,
                                          first_orig_year = 1989,
                                          last_orig_year = 2023)

history <- generate_history_per_claim(data = extended_claims_data,
                                      column = "Cl_payment_cal",
                                      first_orig_year = 1989,
                                      last_orig_year = 2023)

series <- sicr_series(n = 3,
                      large_claims_list = large_claims_list,
                      first_orig_year = 1989,
                      last_orig_year = 2023,
                      pools = pools,
                      indices = indices_xmpl,
                      history = history)
}
