% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sicr.R
\name{sicr_single}
\alias{sicr_single}
\title{Simplified version of sicr}
\usage{
sicr_single(
  dev_year,
  reserve_class,
  last_orig_year,
  pools,
  indices,
  age_shift = NULL,
  mortality = NULL,
  n = 100
)
}
\arguments{
\item{dev_year}{Integer or integer vector. Must be greater or equal to 1.}

\item{reserve_class}{Integer or integer vector. Must be greater or equal to 0.}

\item{last_orig_year}{Last origin year.}

\item{pools}{List as output of \code{\link[=generate_pools]{generate_pools()}}.}

\item{indices}{Dataframe for indexation, see details of \code{\link[=prepare_data]{prepare_data()}}.}

\item{age_shift}{Dataframe, see description of \link{age_shift_xmpl}. Default: NULL \cr
Only necessary if annuities shall be considered.}

\item{mortality}{Dataframe, see description of \link{mortality_xmpl}. Default: NULL \cr
Only necessary if annuities shall be considered.}

\item{n}{Number of simulations. Default: 100}
}
\value{
Dataframe with columns 'Reserve_class', 'Dev_year' and 'Best_estimate'.
}
\description{
This function offers a simplified version of \code{\link[=sicr]{sicr()}}. It can be used to used to
get the gross results for certain development years and reserve classes without building
all the for \code{\link[=sicr]{sicr()}} required objects in advance.
}
\examples{
# this example uses data provided with this package
extended_claims_data <- prepare_data(claims_data = claims_data_xmpl,
                                     indices = indices_xmpl,
                                     threshold = 400000,
                                     first_orig_year = 1989,
                                     last_orig_year = 2023,
                                     expected_year_of_growing_large = 3,
                                     reserve_classes = c(1, 200001, 400001, 700001, 1400001),
                                     pool_of_annuities = pool_of_annuities_xmpl)

pools <- generate_pools(extended_claims_data = extended_claims_data,
                        reserve_classes = c(1, 200001, 400001, 700001, 1400001),
                        years_for_pools = 2014:2023,
                        start_of_tail = 17,
                        end_of_tail = 50,
                        lower_outlier_limit = -Inf,
                        upper_outlier_limit = Inf,
                        pool_of_annuities = pool_of_annuities_xmpl)

sicr_single(dev_year = 1:5,
            reserve_class = 0:3,
            last_orig_year = 2023,
            pools = pools,
            indices = indices_xmpl,
            age_shift = age_shift_xmpl,
            mortality = mortality_xmpl)
}
