% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{skeleton_prediction}
\alias{skeleton_prediction}
\title{Skeleton for payments prediction}
\usage{
skeleton_prediction(first_orig_year, last_orig_year)
}
\arguments{
\item{first_orig_year}{desired first origin year}

\item{last_orig_year}{desired last origin year}
}
\value{
empty prediction matrix
}
\description{
Creates an empty matrix with each historic origin year as row and 250 future calendar years in columns.
}
\examples{
skeleton_prediction(2010, 2015)
}
