% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{skeleton_triangle}
\alias{skeleton_triangle}
\title{Skeleton for history triangle}
\usage{
skeleton_triangle(first_orig_year, last_orig_year)
}
\arguments{
\item{first_orig_year}{desired first origin year}

\item{last_orig_year}{desired last origin year}
}
\value{
empty matrix containing claims history triangle
}
\description{
Creates an empty matrix with each historic origin year as row and as calendar year in columns
}
\examples{
skeleton_triangle(2010, 2015)
}
