% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{tail_factor}
\alias{tail_factor}
\title{Payment transition from first tail year to another development year}
\usage{
tail_factor(dev_year, start_of_tail, end_of_tail)
}
\arguments{
\item{dev_year}{desired development year}

\item{start_of_tail}{development year of tail start}

\item{end_of_tail}{development year of tail end}
}
\value{
numeric factor to apply to payment in first tail year
}
\description{
Vectorized helper function for applying the tail method. Melts down payment from first tail year to desired development year.
}
\details{
As data becomes thin in later development years, for each reserve class all payments between \code{start_of_tail} and \code{end_of_tail} are grouped and therefore scaled to the year \code{start_of_tail} and vice versa by
\deqn{Z_{M+j} = Z_{M+1} \cdot \frac{T - M - j + 1}{T - M}, \quad j = 1, \ldots, T-M}
}
\examples{
tail_factor(25, 17, 50)
tail_factor(17, 17, 50)
tail_factor(50, 17, 50)
}
