% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sicr.R
\name{sicr}
\alias{sicr}
\title{Simulation of large claims per stochastic individual claims reserving}
\usage{
sicr(
  n,
  large_claims_list,
  first_orig_year,
  last_orig_year,
  pools,
  indices,
  history,
  exposure = NULL,
  years_for_ibnr_pools = NULL,
  active_annuities = NULL,
  age_shift = NULL,
  mortality = NULL,
  reinsurance = NULL,
  seed = 12345,
  progress = TRUE,
  summary = TRUE,
  detailed_export = FALSE
)
}
\arguments{
\item{n}{Integer. Desired number of realisations.}

\item{large_claims_list}{Dataframe with one row per known large claim generated by \code{\link[=generate_claims_list]{generate_claims_list()}}.}

\item{first_orig_year}{First origin year with complete history.}

\item{last_orig_year}{Last origin year.}

\item{pools}{List as output of \code{\link[=generate_pools]{generate_pools()}}.}

\item{indices}{Dataframe for indexation, see details of \code{\link[=prepare_data]{prepare_data()}}.}

\item{history}{numeric matrix with one row per claim with historic payments,
generated by \code{\link[=generate_history_per_claim]{generate_history_per_claim()}}. \cr
If generated in other ways, note that colnames must be the origin years. \cr
This matrix is needed to calculate ibnr payments and reinsurance if desired.}

\item{exposure}{Dataframe that must contain one row for each origin year between \code{first_orig_year}
and \code{last_orig_year} and the columns \code{Origin_year} and \code{Exposure}. \cr
Default: NULL. In this case, a constant exposure is assumed.}

\item{years_for_ibnr_pools}{Vector containing the calendar years that shall be used to build the pools. \cr
Default: NULL. In this case, all years are used for building ibnr pools.}

\item{active_annuities}{Dataframe of active annuities, see \link{active_annuities_xmpl}. Default: NULL.\cr
Only necessary if annuities shall be considered.}

\item{age_shift}{Dataframe, see description of \link{age_shift_xmpl}. Default: NULL \cr
Only necessary if annuities shall be considered.}

\item{mortality}{Dataframe, see description of \link{mortality_xmpl}. Default: NULL \cr
Only necessary if annuities shall be considered.}

\item{reinsurance}{Dataframe with reinsurance information, see details of \code{\link[=xl_cashflow]{xl_cashflow()}} or \link{reinsurance_xmpl}. Default: NULL\cr
Only necessary if reinsurance shall be considered.}

\item{seed}{Default: 12345. Setting a seed makes the simulation reproducible.}

\item{progress}{Default: TRUE. Set to FALSE if no progress bar shall be shown during simulation.}

\item{summary}{Default: TRUE. Set to FALSE if no summary shall be shown after the simulation.}

\item{detailed_export}{Default: FALSE. Set to TRUE if the output shall contain more detailed technical information
about the simulation.}
}
\value{
Numeric three-dimensional array with \cr
\itemize{
\item one row per known claim followed by one row per expected ibnr claim
\item 250 columns for future calendar years
\item 7 (or 15 if \code{detailed_export} is true) matrices for the following information:
\enumerate{
\item Sum of gross payments (Gross_sum)
\item Sum of reinsurance payments (Reinsurance_sum)
\item Claim payments (without annuity payments) (Claim_payments)
\item Payments for active annuities (Active_annuities_payments)
\item Payments for expected future annuities (Future_annuities_payments)
\item Reinsurance payments from xl treaties (Reinsurance_xl)
\item Reinsurance payments from quota share treaties (Reinsurance_quota_share)
\item (Reserve class (Reserve_class))
\item (Development year (Dev_year))
\item (Sampled future annuities index 1 (Future_annuities_index1))
\item (Sampled future annuities index 2 (Future_annuities_index2))
\item (Sampled future annuities index 3 (Future_annuities_index3))
\item (Sampled future annuities index 4 (Future_annuities_index4))
\item (Sampled future annuities index 5 (Future_annuities_index5))
\item (Random sampling index (Random_sampling_index))
}
}
}
\description{
This function is the main function of the sicr package. It uses the prepared objects to simulate
future payments for large claims.
}
\examples{
# this example uses data provided with this package
extended_claims_data <- prepare_data(claims_data = claims_data_xmpl,
                                     indices = indices_xmpl,
                                     threshold = 400000,
                                     first_orig_year = 1989,
                                     last_orig_year = 2023,
                                     expected_year_of_growing_large = 3,
                                     reserve_classes = c(1, 200001, 400001, 700001, 1400001),
                                     pool_of_annuities = pool_of_annuities_xmpl)

pools <- generate_pools(extended_claims_data = extended_claims_data,
                        reserve_classes = c(1, 200001, 400001, 700001, 1400001),
                        years_for_pools = 2014:2023,
                        start_of_tail = 17,
                        end_of_tail = 50,
                        lower_outlier_limit = -Inf,
                        upper_outlier_limit = Inf,
                        pool_of_annuities = pool_of_annuities_xmpl)

large_claims_list <- generate_claims_list(extended_claims_data = extended_claims_data,
                                          first_orig_year = 1989,
                                          last_orig_year = 2023)

history <- generate_history_per_claim(data = extended_claims_data,
                                      column = "Cl_payment_cal",
                                      first_orig_year = 1989,
                                      last_orig_year = 2023)
# smallest version
result <- sicr(n = 5,
               large_claims_list = large_claims_list,
               first_orig_year = 1989,
               last_orig_year = 2023,
               pools = pools,
               indices = indices_xmpl,
               history = history)

# more complex version
result <- sicr(n = 5,
               large_claims_list = large_claims_list,
               first_orig_year = 1989,
               last_orig_year = 2023,
               pools = pools,
               indices = indices_xmpl,
               history = history,
               exposure = exposure_xmpl,
               years_for_ibnr_pools = 2014:2023,
               active_annuities = active_annuities_xmpl,
               age_shift = age_shift_xmpl,
               mortality = mortality_xmpl,
               reinsurance = reinsurance_xmpl)
}
